add options to cli mode
