+++ docs/app.js
@@ -1,1 +1,1 @@
-!function(t){var e={};function n(r){if(e[r])return e[r].exports;var i=e[r]={i:r,l:!1,exports:{}};return t[r].call(i.exports,i,i.exports,n),i.l=!0,i.exports}n.m=t,n.c=e,n.d=function(t,e,r){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:r})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var i in t)n.d(r,i,function(e){return t[e]}.bind(null,i));return r},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=3)}([function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});const r=n(4),i=n(2),s=n(5);class a{constructor(t,e,n,i){let a=t.split("#")[0].trim(),o=[];for(;;){let t=a.match(/^[^[]+?\s*=\s*/);if(null===t)break;let e=t[0];(t=a.split(e)).shift(),a=t.join(e),o.push(e.split(/\s*=\s*$/)[0])}this.expr=r.makeAST(a,e),this.original=this.expr,this.type=this.getType(e);for(let t of o)r.Macro.add(t,this,e);this.typed=e,this.etaAllowed=n,this.allowMultipleEdges=i,this.processTex="\\begin{eqnarray*}\n&& ",this.curStep=0,this.root=s.ReductionNode.makeRoot(this.expr,this.typed,this.etaAllowed,this.allowMultipleEdges),this.curNodes=[this.root],this.nextRedexes=void 0}getRedexes(){return this.nextRedexes?this.nextRedexes:this.nextRedexes=this.expr.getRedexes(this.typed,this.etaAllowed,!0).sort(r.Redex.compare)}getLeftMostRedex(){return this.typed?this.getRedexes()[0]||null:this.nextLeftMostRedex?this.nextLeftMostRedex:this.nextLeftMostRedex=this.expr.getLeftMostRedex(this.typed,this.etaAllowed,!0)}reduction(t){if(void 0===t&&!(t=this.getLeftMostRedex()))return null;let e;if(this.expr=t.next,this.nextRedexes=void 0,this.nextLeftMostRedex=void 0,this.processTex+=t.toTexString(),"macro"===t.type?(this.processTex+=" \\\\\n&\\equiv& ",e="-: (macro) = "+this.expr.toString(!0)):(this.processTex+=" \\\\\n&\\longrightarrow_{"+t.getTexRule()+"}& ",e=++this.curStep+": ("+t.rule+") --\x3e "+this.expr.toString(!0)),!this.hasNext()){e+="    (normal form)\n";let t=this.parseChurchNum();null!==t&&(e+="  = "+t+" (as nat)\n");let n=this.parseChurchBool();null!==n&&(e+="  = "+n+" (as bool)\n"),e.slice(0,e.length-1)}return e}deepen(t){if(0===this.curNodes.length)return null;let e=this.curNodes.shift();if(void 0!==t&&e.depth>=t)return this.curNodes.push(e),null;let n=e.visit();for(let t of n.nodes)this.curNodes.push(t);return n}hasNext(){return null!==this.getLeftMostRedex()}hasNodes(){return this.curNodes.length>0}getProofTree(){return"\\begin{prooftree}\n"+this.proofTree+"\\end{prooftree}"}getProcessTex(){return this.processTex+this.expr.toTexString(!0)+(this.hasNext()?"":"\\not\\longrightarrow")+"\n\\end{eqnarray*}"}getType(t){if(!t)return new i.TypeUntyped;i.TypeVariable.maxId=void 0;let e=i.TypeVariable.getNew(),n=this.expr.getEquations([],e,!0),r=n.eqs;this.proofTree=n.proofTree;let s=i.TypeEquation.get(e,i.TypeEquation.solve(r)),a=s.getVariables(),o=[];for(let t of a)i.TypeVariable.contains(o,t)||o.push(t);let l=0;for(let t of o)s.replace(t,i.TypeVariable.getAlphabet(l)),l++;return s}static parseMacroDef(t,e){let n=t.split("#")[0].trim(),i=[];for(;;){let t=n.match(/^[^[]+?\s*=\s*/);if(null===t)break;let e=t[0];(t=n.split(e)).shift(),n=t.join(e),i.push(e.split(/\s*=\s*$/)[0])}if(0===i.length)return null;let s=new a(n,e,void 0,!1);for(let t of i)r.Macro.add(t,s,e);return{names:i,expr:s.expr.toString(!0),type:s.type.toString()}}static fileInput(t,e){let n=t.split("\n"),r=[],i=[];for(let t of n)try{let n=a.parseMacroDef(t,e);null!==n&&i.push(n)}catch(t){r.push(t.toString())}return{defs:i,errs:r}}static getMacroList(t){let e="",n=r.Macro.getMap(t);for(let t in n){let r=n[t];e+="<"+r.name+"> is defined as "+r.expr.toString(!0)+" : "+r.type+"\n"}return e}static getMacroListAsObject(t){return r.Macro.getMap(t)}static clearMacro(t){return r.Macro.clear(t)}static graph2LF(t,e){return s.GraphNode.search(s.GraphNode.parse(t),e)}static lmntal2LF(t,e){return new a(r.parseLMNtal(t).toString(!0),!1,!1,e)}toLMNtal(){return a.nextLinkID=0,this.typed?null:this.original.toLMNtal()}toString(){let t=this.expr.toString(!0)+" : "+this.type;if(!this.hasNext()){t+="    (normal form)\n";let e=this.parseChurchNum();null!==e&&(t+="  = "+e+" (as nat)\n");let n=this.parseChurchBool();null!==n&&(t+="  = "+n+" (as bool)\n"),t=t.slice(0,t.length-1)}return t}getOriginalString(){return this.original+" : "+this.type}parseChurchNum(){return this.expr.parseChurchNum()}parseChurchBool(){return this.expr.parseChurchBool()}static getNewLink(){return"R"+a.nextLinkID++}}e.LambdaFriends=a},function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});class r{constructor(t,e){void 0===e?(this.name="LambdaFriendsError",this.message=t):(this.name=t,this.message=e)}toString(){return this.name+": "+this.message}}e.LambdaFriendsError=r;e.LambdaParseError=class extends r{constructor(t){super("LambdaParseError",t)}};e.SubstitutionError=class extends r{constructor(t){super("SubstitutionError",t)}};e.ReductionError=class extends r{constructor(t){super("ReductionError",t)}};e.MacroError=class extends r{constructor(t){super("MacroError",t)}};e.TypeError=class extends r{constructor(t){super("TypeError",t)}};e.TexError=class extends r{constructor(t){super("TexError",t)}};e.GraphParseError=class extends r{constructor(t){super("GraphParseError",t)}}},function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});const r=n(1);class i{constructor(t){this.className=t}}e.Type=i;class s{constructor(t,e){this.left=t,this.right=e}transform(t,e){if(this.left instanceof a&&this.right instanceof a)return this.left.match(this.right);if(this.left.equals(this.right))return[];if(!(this.left instanceof c))return[new s(this.right,this.left)];if(this.right.contains(this.left))throw new r.TypeError("Illegal type ("+this.right+" contains "+this.left+". Self-application?)");for(var n of t)n.replace(this.left,this.right);for(var n of e)n.replace(this.left,this.right);return[this]}static isEqual(t,e){if(t.length!==e.length)return!1;for(var n of t)if(!s.contains(n,e))return!1;return!0}static contains(t,e){for(var n of e)if(t.equals(n))return!0;return!1}equals(t){return t.left.equals(this.left)&&t.right.equals(this.right)}static get(t,e){for(var n of e)if(n.left.equals(t))return n.right;throw new r.TypeError("Undefined TypeVariable: "+t)}toString(){return this.left+" = "+this.right}replace(t,e){this.left.equals(t)?this.left=e:this.left.replace(t,e),this.right.equals(t)?this.right=e:this.right.replace(t,e)}static solve(t){for(;;){for(var e=[].concat(t),n=[];t.length>0;){var r=t.shift().transform(t,n);n=n.concat(r)}if(t=[].concat(n),s.isEqual(e,n))break}return t}}e.TypeEquation=s;class a extends i{}e.TypeConstructor=a;class o extends a{constructor(){super("TypeInt")}toString(){return"int"}toTexString(){return"{\\rm int}"}equals(t){return t instanceof o}match(t){if(t instanceof o)return[];throw new r.TypeError(this+" and "+t+" are not compatible.")}contains(t){return!1}replace(t,e){}getVariables(){return[]}}e.TypeInt=o;class l extends a{constructor(){super("TypeBool")}toString(){return"bool"}toTexString(){return"{\\rm bool}"}equals(t){return t instanceof l}match(t){if(t instanceof l)return[];throw new r.TypeError(this+" and "+t+" are not compatible.")}contains(t){return!1}replace(t,e){}getVariables(){return[]}}e.TypeBool=l;class h extends a{constructor(t){super("TypeList"),this.content=t}toString(){return"list("+this.content+")"}toTexString(){return"{\\rm list}("+this.content.toTexString()+")"}equals(t){return t instanceof h&&this.content.equals(t.content)}match(t){if(t instanceof h)return[new s(this.content,t.content)];throw new r.TypeError(this+" and "+t+" are not compatible.")}contains(t){return this.content.contains(t)}replace(t,e){this.content.equals(t)?this.content=e:this.content.replace(t,e)}getVariables(){return this.content.getVariables()}}e.TypeList=h;class u extends a{constructor(t,e){super("TypeFunc"),this.left=t,this.right=e}toString(){return(this.left instanceof u?"("+this.left+")":this.left.toString())+" -> "+this.right}toTexString(){return(this.left instanceof u?"("+this.left.toTexString()+")":this.left.toTexString())+" \\rightarrow "+this.right.toTexString()}equals(t){return t instanceof u&&(t.left.equals(this.left)&&t.right.equals(this.right))}match(t){if(t instanceof u)return[new s(this.left,t.left),new s(this.right,t.right)];throw new r.TypeError(this+" and "+t+" are not compatible.")}contains(t){return this.left.contains(t)||this.right.contains(t)}replace(t,e){this.left.equals(t)?this.left=e:this.left.replace(t,e),this.right.equals(t)?this.right=e:this.right.replace(t,e)}getVariables(){return this.left.getVariables().concat(this.right.getVariables())}}e.TypeFunc=u;class c extends i{constructor(t){super("TypeVariable"),this.id=t}toString(){return this.id<0?"'"+c.alphabet[-this.id-1]:"'t"+this.id}toTexString(){return this.id<0?c.texAlphabet[-this.id-1]:"\\tau_{"+this.id+"}"}equals(t){return t instanceof c&&this.id===t.id}static getNew(){return void 0===c.maxId?(c.maxId=0,new c(0)):(c.maxId++,new c(c.maxId))}contains(t){return this.equals(t)}replace(t,e){}getVariables(){return[this]}static getAlphabet(t){return new c(-t-1)}static contains(t,e){for(var n of t)if(n.equals(e))return!0;return!1}}c.alphabet="abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ".split(""),c.texAlphabet=["\\alpha","\\beta","\\gamma","\\delta","\\varepsilon","\\zeta","\\eta","\\theta","\\iota","\\kappa","\\mu","\\nu","\\xi","\\pi","\\rho","\\sigma","\\upsilon","\\phi","\\chi","\\psi","\\omega","\\Gamma","\\Delta","\\Theta","\\Xi","\\Pi","\\Sigma","\\Phi","\\Psi","\\Omega"].concat(c.alphabet),e.TypeVariable=c;class d extends i{constructor(){super("TypeUntyped")}toString(){return"Untyped"}toTexString(){return"{\\rm Untyped}"}equals(t){return t instanceof d}contains(t){return!1}replace(t,e){}getVariables(){return[]}}e.TypeUntyped=d},function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});const r=n(0);let i=cytoscape({container:document.getElementById("graph"),boxSelectionEnabled:!1,autounselectify:!0,style:[{selector:"node",style:{label:"data(label)","background-color":"#11479e"}},{selector:"edge",style:{"target-arrow-shape":"triangle","curve-style":"bezier","target-arrow-color":"#9dbaea",width:3,"line-color":"#9dbaea"}},{selector:".goal",style:{"background-color":"#b3424a"}}]});MicroModal.init({disableScroll:!1,awaitCloseAnimation:!0});let s=void 0,a=!0,o=!1,l=!1,h=void 0,u=document.getElementById("input"),c=document.getElementById("output"),d=document.getElementById("untyped"),f=document.getElementById("typed"),p=document.getElementById("etaEnable"),g=document.getElementById("etaDisable"),m=document.getElementById("multiEdgeEnable"),x=document.getElementById("multiEdgeDisable"),w=document.getElementById("fileInput"),b=new FileReader,T=document.getElementById("clearMacroBtn"),y=document.getElementById("translate"),v=document.getElementById("tabA"),S=document.getElementById("outputBtns"),E=document.getElementById("stepInput"),L=document.getElementById("startGraph"),q=document.getElementById("stopGraph"),R=document.getElementById("imgGraph"),M=document.getElementById("maxDepth"),N=document.getElementById("tabAbtn"),F=document.getElementById("tabBbtn"),k=document.getElementById("tabCbtn"),I=document.getElementById("tabDbtn"),B=document.getElementById("lmnInput"),A=document.getElementById("lmnSubmit"),V=document.getElementById("lmnOutput"),C=document.getElementById("graphInput"),P=document.getElementById("graphSubmit"),U=document.getElementById("graphOutput");w.addEventListener("change",function(t){let e=t.target.files[0];if("text/plain"!==e.type)return alert("プレーンテキストを選択してください"),void(w.value="");b.readAsText(e)}),b.addEventListener("load",function(){let t=r.LambdaFriends.fileInput(b.result,a);tt();let e=document.getElementById("fileInputLog");if(e.textContent=null,t.defs.length>0){let n=document.createElement("div"),r=document.createElement("p");r.innerText="Finally, "+t.defs.length+" macros are successfully added.";let i=document.createElement("ul");i.className="code";for(let e of t.defs){let t=document.createElement("li"),n=[].concat(e.names),r="<"+n.shift()+">";for(;n.length>0;){r+=" = <"+n.shift()+">"}t.innerText=r+" is defined as "+e.expr+" : "+e.type,i.appendChild(t)}n.appendChild(r),n.appendChild(i),e.appendChild(n)}if(t.errs.length>0){let n=document.createElement("div"),r=document.createElement("p");r.innerText="Unfortunately, "+t.errs.length+" macros are rejected due to some errors.";let i=document.createElement("ul");i.className="code";for(let e of t.errs){let t=document.createElement("li");t.innerText=e,i.appendChild(t)}n.appendChild(r),n.appendChild(i),e.appendChild(n)}MicroModal.show("modal-1",{debugMode:!0,disableScroll:!0,awaitCloseAnimation:!0})}),d.onclick=function(){d.className="btn btn-primary",f.className="btn btn-default",a=!1,p.disabled=!1,g.disabled=!1,m.disabled=!1,x.disabled=!1,tt()},f.onclick=function(){f.className="btn btn-primary",d.className="btn btn-default",a=!0,p.disabled=!0,g.disabled=!0,m.disabled=!0,x.disabled=!0,tt()},p.onclick=function(){p.className="btn btn-primary",g.className="btn btn-default",o=!0},g.onclick=function(){g.className="btn btn-primary",p.className="btn btn-default",o=!1},m.onclick=function(){m.className="btn btn-primary",x.className="btn btn-default",l=!0},x.onclick=function(){x.className="btn btn-primary",m.className="btn btn-default",l=!1},T.onclick=function(){r.LambdaFriends.clearMacro(a),tt()},E.addEventListener("change",function(){let t=parseInt(E.value);s=isNaN(t)?void 0:t});let $=[],j=0,z=[""],O=function(){let t=u.value;if(""!==t||void 0===h){$.unshift(t),j=0,(z=[].concat($)).unshift(""),t=(t=t.split("#")[0]).trim();try{let e=r.LambdaFriends.parseMacroDef(t,a);if(null===e)h=new r.LambdaFriends(t,a,o,l),H=0,i.remove("*"),i.add({group:"nodes",data:{id:""+h.root.id,label:h.root.toString()},classes:h.root.isNormalForm?"goal":""}),ot(),W(h.toString()),a&&st(),at();else{let t=[].concat(e.names),n="<"+t.shift()+">";for(;t.length>0;){n+=" = <"+t.shift()+">"}W(n+=" is defined as "+e.expr+" : "+e.type),S.textContent=null}nt()}catch(t){W(t.toString()),console.log(t),S.textContent=null}tt(),u.value=""}else X?Z():st()};document.getElementById("submit").onclick=O,document.getElementById("input").onkeydown=function(t){13===t.keyCode?(O(),t.preventDefault()):38===t.keyCode?(j<z.length-1&&(z[j]=u.value,j++,u.value=z[j]),t.preventDefault()):40===t.keyCode&&(j>0&&(z[j]=u.value,j--,u.value=z[j]),t.preventDefault())};let _,D=!1,H=0,G=!1;function Z(){if(G)return;if(ot(),void 0===h)return;D=!1,G=!0,H+=_||10;let t=()=>setTimeout(()=>{if(D)return ot(),void(G=!1);let e=h.deepen(H);if(null===e)return ot(),void(G=!1);let n=[];for(let t of e.nodes)n.push({group:"nodes",data:{id:""+t.id,label:""+t.toString()},classes:t.isNormalForm?"goal":""});for(let t of e.edges)n.push({group:"edges",data:{source:t.from.id.toString(),target:t.to.id.toString()}});ot(),i.add(n),t(),ot()},1);t()}L.onclick=Z,q.onclick=function(){D=!0},R.onclick=function(){window.open(i.png())},M.addEventListener("change",function(){let t=parseInt(M.value);_=isNaN(t)?void 0:t});let X=!1;function K(){let t=B.value;try{let e=r.LambdaFriends.lmntal2LF(t,l);V.innerText="Found: "+e.expr.toString(!0)}catch(t){V.innerText=t.toString()}}function J(){let t=C.value;try{let e=r.LambdaFriends.graph2LF(t,l);U.innerText=null===e?"Not Found":"Found: "+e.expr.toString(!0)}catch(t){U.innerText=t.toString()}}N.addEventListener("click",()=>{X=!1}),F.addEventListener("click",()=>{X=!1}),k.addEventListener("click",()=>{X=!1}),I.addEventListener("click",()=>{setTimeout(ot,10),X=!0}),B.onkeydown=function(t){13===t.keyCode&&(K(),t.preventDefault())},A.onclick=K,C.onkeydown=function(t){13===t.keyCode&&(J(),t.preventDefault())},P.onclick=J;let Q="";function W(t){Q=t+"\n",c.innerHTML=et(Q)}function Y(t){Q+=t+"\n",c.innerHTML=et(Q)}function tt(){let t=document.getElementById("macroList");t.innerHTML="";let e=r.LambdaFriends.getMacroListAsObject(a);for(let n in e){let r=e[n];t.innerHTML+="<tr><th>"+et(r.name)+"</th><td>"+et(r.expr.toString(!0))+"</td><td>"+et(r.type.toString())+"</td></tr>"}}function et(t){return t.replace(/[&'`"<> \n]/g,function(t){return{"&":"&amp;","'":"&#x27;","`":"&#x60;",'"':"&quot;","<":"&lt;",">":"&gt;"," ":"&nbsp;","\n":"<br>"}[t]})}function nt(){y.textContent=null,void 0!==h&&(y.appendChild(rt("これまでの簡約過程",h.getProcessTex())),a?y.appendChild(rt("型付けの証明木",h.getProofTree())):y.appendChild(rt("LMNtalコード",h.toLMNtal())))}function rt(t,e){let n=document.createElement("p"),r=document.createElement("button"),i=document.createElement("span"),s=document.createElement("div"),a=document.createElement("p");s.classList.add("code"),n.appendChild(r),n.appendChild(i),s.appendChild(a),a.innerText=e,i.innerText=t,r.type="button",r.className="btn btn-default btn-sm",r.innerText="クリップボードにコピー",r.onclick=function(){let t=document.getSelection();t.selectAllChildren(s),document.execCommand("copy"),t.collapse(document.body,0)};let o=document.createElement("div");return o.appendChild(n),o.appendChild(s),o}let it=!1;function st(){if(it)return;it=!0,S.textContent=null;let t=e=>setTimeout(()=>{if(0===e||!h.hasNext())return at(),v.scrollTop=c.offsetHeight-15,void(it=!1);Y(h.reduction()),v.scrollTop=c.offsetHeight-15,nt(),t(e-1)},1);t(void 0===s?100:s)}function at(){if(!h.hasNext())return void(S.textContent=null);let t=document.createElement("button");if(t.type="button",t.className="btn btn-default btn-sm",t.innerText="最左簡約を続ける",t.onclick=st,S.textContent=null,S.appendChild(t),a)return;let e=document.createElement("span");e.innerText="または、以下から簡約基を選ぶ：",S.appendChild(e);let n=document.createElement("div");n.className="list-group",S.appendChild(n);let r=h.getRedexes();for(let t of r){let e=document.createElement("button");e.className="list-group-item code",e.innerHTML=t.toHTMLString(),e.onclick=function(){Y(h.reduction(t)),at(),v.scrollTop=c.offsetHeight-15,nt()},n.appendChild(e)}}function ot(){i.resize(),i.elements().makeLayout({name:"dagre",nodeSpacing:5,animate:!0,randomize:!1,maxSimulationTime:1500}).run()}d.onclick(null),g.onclick(null),x.onclick(null),tt()},function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});const r=n(2),i=n(1),s=n(0);function a(t,e){let n=null;for(;t.length>0;){let r=t.shift();if(r instanceof w||r instanceof v||r instanceof S)n=null===n?r:new L(n,r);else switch(r.name){case"\\":case"¥":case"λ":return null===n?E.parse(t,e):new L(n,E.parse(t,e));case"(":{let r=[],s=1;for(;;){if(0==t.length)throw new i.LambdaParseError("Too many LPAREN '('");let e=t.shift();if("("===e.name?s++:")"===e.name&&s--,0==s)break;r.push(e)}let o=a(r,e);n=null===n?o:new L(n,o);break}default:if(e)switch(r.name){case"if":return R.parse(t,e);case"let":return M.parse(t,e);case"case":return N.parse(t,e);case"fix":return F.parse(t,e);case":":{let r=t.shift();if(":"!==r.name)throw new i.LambdaParseError("Unexpected token: '"+r+"'");return new q(n,a(t,e))}}if(null===r.name.match(/^[A-Za-z]$/))throw new i.LambdaParseError("Unexpected token: '"+r+"'");n=null===n?new x(r.name):new L(n,new x(r.name))}}if(null===n)throw new i.LambdaParseError("No contents in Expression");return n}function o(t,e){return a(function(t,e){let n=t.split(""),r=[];for(;n.length>0;){let t=n.shift().trim();if(""===t);else if("<"===t){let s="";for(;;){if(0==n.length)throw new i.LambdaParseError("Too many LANGLE '<'");if(">"===(t=n.shift()))break;s+=t}r.push(S.get(s,e))}else if(e&&"["===t){let e="";for(;;){if(0==n.length)throw new i.LambdaParseError("Too many LBRACKET '['");if("]"===(t=n.shift()))break;e+=t}let s=null;switch(e){case"nil":s=new v;break;case"false":case"true":s=new T("true"===e);break;case"if":case"then":case"else":case"let":case"in":case"case":case"of":case"fix":s=new m(e);break;default:s=null!==e.match(/^\d+$|^-\d+$/)?new b(parseInt(e)):new y(e)}if(null===s)throw new i.LambdaParseError("Unknown Const: ["+e+"]");r.push(s)}else r.push(new m(t))}return r}(t,e),e)}function l(t){return t.replace(/[&'`"<>]/g,function(t){return{"&":"&amp;","'":"&#x27;","`":"&#x60;",'"':"&quot;","<":"&lt;",">":"&gt;"}[t]})}e.makeAST=o,e.parseLMNtal=function(t){return function t(n,r,s){let a=(n=n.trim()).match(/^.+?(?=\()/);if(null===a){let t=s[n];if(void 0===t)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. Unknown Token: "+n);return new x(t)}let o=a[0].trim(),l=e(n.match(/\(.+$/)[0]);switch(o){case"lambda":{if(2!==l.length)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. lambda(X,A) should have 2 args.");let n=x.getNew(r),a=function t(n){let r=(n=n.trim()).match(/^.+?(?=\()/);if(null===r)return"rm"===n?[]:[n];let s=r[0].trim(),a=e(n.match(/\(.+$/)[0]);if("rm"===s){if(0!==a.length)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. rm() should have 0 args.");return[]}if("cp"===s){if(2!==a.length)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. cp(A,B) should have 2 args.");return t(a[0].trim()).concat(t(a[1].trim()))}throw new i.LambdaParseError("Malformed LMNtal Lambda Term. Unexpected atom name: "+s)}(l[0]);for(let t of a)s[t]=n.name;return new E(n,t(l[1],r.concat(n),Object.assign({},s)))}case"apply":if(2!==l.length)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. apply(A,B) should have 2 args.");return new L(t(l[0],[].concat(r),Object.assign({},s)),t(l[1],[].concat(r),Object.assign({},s)));case"fv":if(1!==l.length)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. fv(X) should have 1 arg.");return S.get(l[0],!1);default:throw new i.LambdaParseError("Malformed LMNtal Lambda Term. Unexpected atom name: "+o)}}(t,[],{});function e(t){if("("!==(t=t.trim())[0]||")"!==t[t.length-1])throw new i.LambdaParseError("Malformed LMNtal Lambda Term. Invalid Parentheses.");let e=0,n="",r=[];for(let i=1;i<t.length-1;i++){let s=t[i];"("===s?e++:")"===s&&e--,0===e&&","===s?(r.push(n),n=""):n+=s}if(0!==e)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. Invalid Parentheses.");if(r.push(n),1===r.length){let t=r[0].trim();return""===t?[]:[t]}let s=[];for(let t of r)s.push(t.trim());return s}},e.makeTerms=function(t){return function t(e,n){let r=[].concat(e);if(0===n)return r;let i=t(e,n-1);for(let t=0;t<i.length;t++)for(let e=0;e<i.length;e++)r.push(new L(i[t],i[e]));let s=x.getNew(e),a=t(e.concat(s),n-1);for(let t of a)r.push(new E(s,t));return r}([],t)};class h{constructor(t){this.left="",this.right="",this.texLeft="",this.texRight="",this.type=t}addLeft(t){this.left=t+this.left}addRight(t){this.right+=t}addTexLeft(t){this.texLeft=t+this.texLeft}addTexRight(t){this.texRight+=t}static makeNext(t,e,n,r,i,s){let a=[].concat(t);for(let t of a)t.next=s(t.next),t.addLeft(e),t.addRight(n),t.addTexLeft(r),t.addTexRight(i);return a}getName(){return this.type}getPos(){return this.left.length}static compare(t,e){let n=t.getPos(),r=e.getPos();return n===r?0:n-r}}e.Redex=h;class u extends h{constructor(t){super("beta"),this.content=t,this.la=t.left,this.next=this.la.expr.substitute(this.la.boundvar,t.right),this.arg=t.right,this.rule="beta"}toString(){let t=[],e=this.la.expr;for(;e instanceof E;)t.push(e.boundvar.toString(!1)),e=e.expr;let n=t.join("")+"."+e.toString(!0);return this.left+"(\\["+this.la.boundvar.toString(!1)+"]"+n+")["+this.arg.toString(!1)+"]"+this.right}toTexString(){let t=[],e=this.la.expr;for(;e instanceof E;)t.push(e.boundvar.toTexString(!1)),e=e.expr;let n=t.join("")+"."+e.toTexString(!0);return this.texLeft+"(\\strut \\lambda{\\underline{"+this.la.boundvar.toTexString(!1)+"}}"+n+")\\underline{\\strut "+this.arg.toTexString(!1)+"}"+this.texRight}toHTMLString(){let t=[],e=this.la.expr;for(;e instanceof E;)t.push(e.boundvar.toString(!1)),e=e.expr;let n=t.join("")+"."+e.toString(!0);return l(this.left)+'(\\<span class="lf-beta lf-boundvar">'+l(this.la.boundvar.toString(!1))+"</span>"+l(n)+')<span class="lf-beta lf-arg">'+l(this.arg.toString(!1))+"</span>"+l(this.right)}getTexRule(){return"\\beta"}}class c extends h{constructor(t){super("eta"),this.content=t,this.app=t.expr,this.next=this.app.left,this.rule="eta"}toString(){return this.left+"["+this.content+"]"+this.right}toTexString(){return this.texLeft+"\\underline{\\strut "+this.content.toTexString(this.left+this.right==="")+"}"+this.texRight}toHTMLString(){return l(this.left)+'<span class="lf-eta">('+l(this.content.toString(this.left+this.right===""))+")</span>"+l(this.right)}getTexRule(){return"\\eta"}}class d extends h{constructor(t){super("macro"),this.content=t,this.next=t.expr,this.rule="macro"}toString(){return this.left+"[<"+this.content.name+">]"+this.right}toTexString(){return this.texLeft+"\\underline{\\strut "+this.content.toTexString(!1)+"}"+this.texRight}toHTMLString(){return l(this.left)+'<span class="lf-macro">&lt;'+l(this.content.name)+"&gt;</span>"+l(this.right)}getTexRule(){return"{\\rm m}"}}class f extends h{constructor(t,e,n){super("typed"),this.content=t,this.next=e,this.rule=n}toString(){return this.left+"["+this.content.toString(this.left+this.right==="")+"]"+this.right}toTexString(){return this.texLeft+"\\underline{\\strut "+this.content.toTexString(!1)+"}"+this.texRight}toHTMLString(){return l(this.left)+'<span class="lf-typed">'+l(this.content.toString(this.left+this.right===""))+"</span>"+l(this.right)}getTexRule(){return"{\\rm ("+this.rule+")}"}}class p{constructor(t,e){this.eqs=t,this.proofTree=e}}class g{constructor(t){this.className=t}parseChurchNum(){if(!(this instanceof E))return null;const t=this.boundvar;let e=this.expr,n=0;if(!(e instanceof E))return null;const r=e.boundvar;for(e=e.expr;e instanceof L;){if(n++,!e.left.equals(t))return null;e=e.right}return e.equals(r)?n:null}parseChurchBool(){const t=new E(new x("x"),new E(new x("y"),new x("x"))),e=new E(new x("x"),new E(new x("y"),new x("y")));return!!this.equalsAlpha(t)||!this.equalsAlpha(e)&&null}toLMNtal(){throw new i.TypeError("Expression '"+this+"' cannot be converted into LMNtal (untyped only).")}}e.Expression=g;class m extends g{constructor(t,e){super(void 0===e?"Symbol":e),this.name=t}equals(t){return t instanceof m&&t.className===this.className&&t.name===this.name}equalsAlpha(t){return t instanceof m&&t.className===this.className&&t.name===this.name}toString(t){return this.name}toTexString(t){throw new i.TexError("class Symbol does not have tex string")}getFV(){return this.freevars}substitute(t,e){throw new i.SubstitutionError("Undefined Substitution")}getEquations(t,e,n){throw new i.TypeError("Undefined Type")}getRedexes(t,e,n){throw new i.ReductionError("Symbols must not appear in parsed Expression")}getLeftMostRedex(t,e,n){throw new i.ReductionError("Symbols must not appear in parsed Expression")}extractMacros(){throw new i.ReductionError("Symbols must not appear in parsed Expression")}}class x extends m{constructor(t){super(t,"Variable"),this.freevars=[this]}substitute(t,e){return this.equals(t)?e:this}getEquations(t,e,n){for(let n of t)if(n.equals(this)){let i="\\AxiomC{}\n";return i+="\\RightLabel{\\scriptsize(var)}\n",i+="\\UnaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.name+" : "+e.toTexString()+" $}\n",new p([new r.TypeEquation(n.type,e)],i)}throw new i.TypeError("free variable is not allowed: "+this)}toTexString(t){return this.name}static union(t,e,n){if(void 0===n){let n=[];for(let e of t)n.push(e);for(let r of x.dif(e,t))n.push(r);return n}return x.union(x.union(t,e),n)}static dif(t,e){let n=[];for(let r of t)x.contains(e,r)||n.push(r);return n}static contains(t,e){for(let n of t)if(n.equals(e))return!0;return!1}static gammaToTexString(t){if(0===t.length)return"";let e=t[0].name+" : "+t[0].type.toTexString();for(let n=1;n<t.length;n++)e+=",~"+t[n].name+" : "+t[n].type.toTexString();return e}static getNew(t){let e="abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ".split("");for(let n of e){let e=new x(n);if(!x.contains(t,e))return e}throw new i.SubstitutionError("No more Variables available")}getRedexes(t,e,n){return[]}getLeftMostRedex(t,e,n){return null}extractMacros(){return this}toLMNtal(){return"fv("+this.name+")"}}class w extends m{constructor(t,e){super(t,e),this.freevars=[]}substitute(t,e){return this}getEquations(t,e,n){let i="\\AxiomC{}\n";return i+="\\RightLabel{\\scriptsize(con)}\n",i+="\\UnaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p([new r.TypeEquation(this.type,e)],i)}toString(t){return"["+this.name+"]"}toTexString(t){return this.name+"^{"+this.type.toTexString()+"}"}getRedexes(t,e,n){if(t)return[];throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className)}getLeftMostRedex(t,e,n){if(t)return null;throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className)}extractMacros(){return this}}class b extends w{constructor(t){super(t.toString(),"ConstInt"),this.value=t,this.type=new r.TypeInt}}class T extends w{constructor(t){super(t.toString(),"ConstBool"),this.value=t,this.type=new r.TypeBool}}class y extends w{constructor(t){switch(super(t,"ConstOp"),t){case"+":this.value=((t,e)=>new b(t.value+e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeInt));break;case"-":this.value=((t,e)=>new b(t.value-e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeInt));break;case"*":this.value=((t,e)=>new b(t.value*e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeInt));break;case"/":this.value=((t,e)=>{if(0===e.value)throw new i.ReductionError("Dividing by '0' is not allowed");return new b(Math.floor(t.value/e.value))}),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeInt));break;case"%":this.value=((t,e)=>{if(0===e.value)throw new i.ReductionError("Dividing by '0' is not allowed");return new b(t.value-4*Math.floor(t.value/e.value))}),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeInt));break;case"<":this.value=((t,e)=>new T(t.value<e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeBool));break;case">":this.value=((t,e)=>new T(t.value>e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeBool));break;case"<=":this.value=((t,e)=>new T(t.value<=e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeBool));break;case">=":this.value=((t,e)=>new T(t.value>=e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeBool));break;case"==":this.value=((t,e)=>new T(t.value==e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeBool));break;case"!=":this.value=((t,e)=>new T(t.value!=e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeBool));break;case"eq":case"eq":this.value=((t,e)=>new T(t.value==e.value)),this.type=new r.TypeFunc(new r.TypeBool,new r.TypeFunc(new r.TypeBool,new r.TypeBool));break;case"xor":this.value=((t,e)=>new T(t.value!=e.value)),this.type=new r.TypeFunc(new r.TypeBool,new r.TypeFunc(new r.TypeBool,new r.TypeBool));break;case"or":this.value=((t,e)=>new T(t.value||e.value)),this.type=new r.TypeFunc(new r.TypeBool,new r.TypeFunc(new r.TypeBool,new r.TypeBool));break;case"and":this.value=((t,e)=>new T(t.value&&e.value)),this.type=new r.TypeFunc(new r.TypeBool,new r.TypeFunc(new r.TypeBool,new r.TypeBool));break;case"nor":this.value=((t,e)=>new T(!(t.value||e.value))),this.type=new r.TypeFunc(new r.TypeBool,new r.TypeFunc(new r.TypeBool,new r.TypeBool));break;case"nand":this.value=((t,e)=>new T(!(t.value&&e.value))),this.type=new r.TypeFunc(new r.TypeBool,new r.TypeFunc(new r.TypeBool,new r.TypeBool));break;default:throw new i.LambdaParseError("Undefined function: "+t)}}}class v extends m{substitute(t,e){return this}constructor(){super("nil","Nil"),this.freevars=[]}getEquations(t,e,n){let i=r.TypeVariable.getNew(),s=new r.TypeList(i),a="\\AxiomC{}\n";return a+="\\RightLabel{\\scriptsize(nil)}\n",a+="\\UnaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.name+" : "+s.toTexString()+" $}\n",new p([new r.TypeEquation(e,s)],a)}toString(t){return"["+this.name+"]"}toTexString(t){return"{\\rm "+this.name+"}"}getRedexes(t,e,n){if(t)return[];throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className)}getLeftMostRedex(t,e,n){if(t)return null;throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className)}extractMacros(){return this}}class S extends m{constructor(t,e,n,r){super(t,"Macro"),this.freevars=[],this.expr=e,this.typed=n,this.type=r}static add(t,e,n){let r=e.expr;if(!/^[a-zA-Z0-9!?]+$/.test(t))throw new i.MacroError("<"+t+"> cannot be used as a name of macro. Available characters: [a-zA-Z0-9!?]");let s=n?S.map:S.mapUntyped;if(0!==r.getFV().length)throw new i.MacroError("<"+t+"> contains free variables: "+r.getFV());let a=new S(t,r,n,e.type);return s[t]=a,s[t]}static clear(t){t?S.map={}:S.mapUntyped={}}static get(t,e){let n;return void 0===(n=e?S.map[t]:S.mapUntyped[t])?null!==t.match(/^\d+$/)?new S(t,function(t,e){let n=0===t?"z":"sz";for(let e=1;e<t;e++)n="s("+n+")";return o("\\sz."+n,e)}(parseInt(t),e),e,void 0):new S(t,"true"==t?o("\\xy.x",e):"false"==t?o("\\xy.y",e):void 0,e,void 0):new S(t,n.expr,e,n.type)}static getMap(t){return Object.assign({},t?S.map:S.mapUntyped)}substitute(t,e){return this}toString(t){return"<"+this.name+">"}equalsAlpha(t){return void 0===this.expr?this.equals(t):this.expr.equalsAlpha(t)}getEquations(t,e,n){if(void 0===this.expr)throw new i.TypeError(this+" is undefined.");return this.expr.getEquations(t,e,n)}toTexString(t){return"\\,\\overline{\\bf "+this.name+"}\\,"}getRedexes(t,e,n){let r=S.get(this.name,t);return void 0===r.expr?[]:[new d(r)]}getLeftMostRedex(t,e,n){let r=S.get(this.name,t);return void 0===r.expr?null:new d(r)}extractMacros(){return void 0===this.expr?this:this.expr.extractMacros()}toLMNtal(){return void 0===this.expr?"fv("+this.name+")":this.expr.toLMNtal()}}S.map={},S.mapUntyped={},e.Macro=S;class E extends g{constructor(t,e){super("LambdaAbstraction"),this.freevars=void 0,this.boundvar=t,this.expr=e}static parse(t,e){let n=[];for(;t.length>0;){let r=t.shift();if("."===r.name){let r=a(t,e);for(;n.length>0;)r=new E(n.pop(),r);return r}if(null===r.name.match(/^[A-Za-z]$/))throw new i.LambdaParseError("Unexpected token: '"+r+"'");n.push(new x(r.name))}throw new i.LambdaParseError("'.' is needed")}toString(t){let e=[this.boundvar],n=this.expr;for(;n instanceof E;)e.push(n.boundvar),n=n.expr;let r="\\"+e.join("")+"."+n.toString(!0);return t||(r="("+r+")"),r}getFV(){return void 0!==this.freevars?this.freevars:(this.freevars=[],this.freevars=x.dif(this.expr.getFV(),[this.boundvar]))}substitute(t,e){if(this.boundvar.equals(t))return this;if(x.contains(e.getFV(),this.boundvar)){let n=x.union(this.expr.getFV(),e.getFV()),r=x.getNew(n);return new E(r,this.expr.substitute(this.boundvar,r)).substitute(t,e)}return new E(this.boundvar,this.expr.substitute(t,e))}equals(t){return t instanceof E&&t.boundvar.equals(this.boundvar)&&t.expr.equals(this.expr)}equalsAlpha(t){if(!(t instanceof E))return!1;if(this.equals(t))return!0;let e=this.boundvar,n=this.expr,r=t.boundvar,i=t.expr;return x.contains(n.getFV(),r)?i.equalsAlpha(n):i.equalsAlpha(n.substitute(e,r))}getEquations(t,e,n){let i=r.TypeVariable.getNew(),s=r.TypeVariable.getNew();this.boundvar.type=s;let a=this.expr.getEquations(t.concat(this.boundvar),i,!0),o=a.proofTree;return o+="\\RightLabel{\\scriptsize(abs)}\n",o+="\\UnaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p(a.eqs.concat(new r.TypeEquation(e,new r.TypeFunc(s,i))),o)}toTexString(t){let e=[this.boundvar.toTexString(!1)],n=this.expr;for(;n instanceof E;)e.push(n.boundvar.toTexString(!1)),n=n.expr;let r="\\lambda "+e.join("")+"."+n.toTexString(!0);return t||(r="("+r+")"),r}isEtaRedex(){return this.expr instanceof L&&this.expr.right.equals(this.boundvar)&&!x.contains(this.expr.left.getFV(),this.boundvar)}getRedexes(t,e,n){if(t)return[];let r=[this.boundvar],i=this.expr;for(;i instanceof E;)r.push(i.boundvar),i=i.expr;let s="",a="";n||(s="(",a=")");let o=h.makeNext(i.getRedexes(!1,e,!0),s+"\\"+r.join("")+".",a,s+"\\lambda{"+r.join("")+"}.",a,t=>{let e=[].concat(r),n=t;for(;e.length>0;){let t=e.pop();n=new E(t,n)}return n});return void 0===e&&(console.error("etaAllowed is undefined."),e=!1),e&&this.isEtaRedex()&&o.push(new c(this)),o}getLeftMostRedex(t,e,n){if(t)return null;if(void 0===e&&(console.error("etaAllowed is undefined."),e=!1),e&&this.isEtaRedex())return new c(this);let r=[this.boundvar],i=this.expr;for(;i instanceof E;)r.push(i.boundvar),i=i.expr;let s="",a="";n||(s="(",a=")");let o=i.getLeftMostRedex(!1,e,!0);return null===o?null:(o.next=(t=>{let e=[].concat(r),n=t;for(;e.length>0;){let t=e.pop();n=new E(t,n)}return n})(o.next),o.addLeft(s+"\\"+r.join("")+"."),o.addRight(a),o.addTexLeft(s+"\\lambda{"+r.join("")+"}."),o.addTexRight(a),o)}extractMacros(){return new E(this.boundvar,this.expr.extractMacros())}toLMNtal(){let t=this.expr.toLMNtal().split("fv("+this.boundvar.name+")"),e=t[0],n=[];for(let r=1;r<t.length;r++){let i=s.LambdaFriends.getNewLink();n.push(i),e+=i+t[r]}return"lambda("+function t(e){switch(e.length){case 0:return"rm";case 1:return e[0];case 2:return"cp("+e[0]+","+e[1]+")";default:return"cp("+e.shift()+","+t(e)+")"}}(n)+","+e+")"}}class L extends g{constructor(t,e){super("Application"),this.left=t,this.right=e}isBetaRedex(){return this.left instanceof E}toString(t){let e=this.left.toString(this.left instanceof L)+this.right.toString(!1);return t||(e="("+e+")"),e}getFV(){return void 0===this.freevars?this.freevars=x.union(this.left.getFV(),this.right.getFV()):this.freevars}substitute(t,e){return new L(this.left.substitute(t,e),this.right.substitute(t,e))}equals(t){return t instanceof L&&t.left.equals(this.left)&&t.right.equals(this.right)}equalsAlpha(t){return t instanceof L&&t.left.equalsAlpha(this.left)&&t.right.equalsAlpha(this.right)}getEquations(t,e,n){let i=r.TypeVariable.getNew(),s=this.left.getEquations(t,new r.TypeFunc(i,e),!1),a=this.right.getEquations(t,i,!1),o=s.proofTree+a.proofTree;return o+="\\RightLabel{\\scriptsize(app)}\n",o+="\\BinaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p(s.eqs.concat(a.eqs),o)}toTexString(t){let e=this.left.toTexString(this.left instanceof L)+this.right.toTexString(!1);return t||(e="("+e+")"),e}getRedexes(t,e,n){if(t){let t=n?"":"(",e=n?"":")";if(this.left instanceof E)return[new f(this,this.left.expr.substitute(this.left.boundvar,this.right),"app2")];if(this.left instanceof L&&this.left.left instanceof y&&this.left.right instanceof w){let n=this.left.left,s=this.left.right,a=this.right;if(a instanceof w){if(n.type.left.equals(s.type)&&n.type.right instanceof r.TypeFunc&&n.type.right.left.equals(a.type))return[new f(this,n.value(s,a),"app5")];throw new i.ReductionError(n.type+" cannot handle "+s.type+" and "+a.type+" as arguments")}return h.makeNext(a.getRedexes(!0,!1,!1),t+n.toString(!1)+s.toString(!1),e,t+n.toTexString(!1)+s.toTexString(!1),e,t=>new L(new L(n,s),t))}return this.left instanceof y?h.makeNext(this.right.getRedexes(!0,!1,!1),t+this.left.toString(!1),e,t+this.left.toTexString(!1),e,t=>new L(this.left,t)):h.makeNext(this.left.getRedexes(!0,!1,!1),t,e+this.right.toString(!1),t,e+this.right.toTexString(!1),t=>new L(t,this.right))}{let t=[this],r=[""],i=[""];for(;;){let e=t[t.length-1];if(r.push(e.right.toString(!1)+r[r.length-1]),i.push(e.right.toTexString(!1)+r[i.length-1]),!(e.left instanceof L))break;t.push(e.left)}let s=t[t.length-1].left.getRedexes(!1,e,!1);for(;t.length>0;){let n=t.pop(),a=h.makeNext(s,"",n.right.toString(!1),"",n.right.toTexString(!1),t=>new L(t,n.right)),o=n.left.toString(!0),l=h.makeNext(n.right.getRedexes(!1,e,!1),o,"",n.left.toTexString(!1),"",t=>new L(n.left,t));s=a.concat(l),r.pop(),i.pop(),n.isBetaRedex()&&s.push(new u(n))}return n||(s=h.makeNext(s,"(",")","(",")",t=>t)),s}}getLeftMostRedex(t,e,n){if(t){let t=n?"":"(",e=n?"":")";if(this.left instanceof E)return new f(this,this.left.expr.substitute(this.left.boundvar,this.right),"app2");if(this.left instanceof L&&this.left.left instanceof y&&this.left.right instanceof w){let n=this.left.left,s=this.left.right,a=this.right;if(a instanceof w){if(n.type.left.equals(s.type)&&n.type.right instanceof r.TypeFunc&&n.type.right.left.equals(a.type))return new f(this,n.value(s,a),"app5");throw new i.ReductionError(n.type+" cannot handle "+s.type+" and "+a.type+" as arguments")}{let r=a.getLeftMostRedex(!0,!1,!1);return null===r?null:(r.next=new L(new L(n,s),r.next),r.addLeft(t+n.toString(!1)+s.toString(!1)),r.addRight(e),r.addTexLeft(t+n.toTexString(!1)+s.toTexString(!1)),r.addTexRight(e),r)}}if(this.left instanceof y){let n=this.right.getLeftMostRedex(!0,!1,!1);return null===n?null:(n.next=new L(this.left,n.next),n.addLeft(t+this.left.toString(!1)),n.addRight(e),n.addTexLeft(t+this.left.toTexString(!1)),n.addTexRight(e),n)}{let n=this.right.getLeftMostRedex(!0,!1,!1);return null===n?null:(n.next=new L(this.left,n.next),n.addLeft(t),n.addRight(e+this.right.toString(!1)),n.addTexLeft(t),n.addTexRight(e+this.right.toTexString(!1)),n)}}{let t=[this],r=[""],i=[""];for(;;){let e=t[t.length-1];if(r.push(e.right.toString(!1)+r[r.length-1]),i.push(e.right.toTexString(!1)+r[i.length-1]),!(e.left instanceof L))break;t.push(e.left)}let s=t[t.length-1].left.getLeftMostRedex(!1,e,!1);for(;t.length>0;){let n=t.pop();if(n.isBetaRedex())s=new u(n);else if(null===s){if(null!==(s=n.right.getLeftMostRedex(!1,e,!1))){let t=n.left.toString(!0);s.next=new L(n.left,s.next),s.addLeft(t),s.addRight(""),s.addTexLeft(n.left.toTexString(!1)),s.addTexRight("")}}else s.next=new L(s.next,n.right),s.addLeft(""),s.addRight(n.right.toString(!1)),s.addTexLeft(""),s.addTexRight(n.right.toTexString(!1));r.pop(),i.pop()}return null===s?null:(n||(s.addLeft("("),s.addRight(")"),s.addTexLeft("("),s.addTexRight(")")),s)}}extractMacros(){return new L(this.left.extractMacros(),this.right.extractMacros())}toLMNtal(){return"apply("+this.left.toLMNtal()+","+this.right.toLMNtal()+")"}}class q extends g{constructor(t,e){super("List"),this.head=t,this.tail=e}toString(t){let e=this.head.toString(!1)+"::"+this.tail.toString(!0);return t||(e="("+e+")"),e}getFV(){return void 0===this.freevars?this.freevars=x.union(this.head.getFV(),this.tail.getFV()):this.freevars}substitute(t,e){return new q(this.head.substitute(t,e),this.tail.substitute(t,e))}equals(t){return t instanceof q&&t.head.equals(this.head)&&t.tail.equals(this.tail)}equalsAlpha(t){return t instanceof q&&t.head.equalsAlpha(this.head)&&t.tail.equalsAlpha(this.tail)}getEquations(t,e,n){let i=r.TypeVariable.getNew(),s=new r.TypeList(i),a=this.head.getEquations(t,i,!1),o=this.tail.getEquations(t,s,!1),l=a.proofTree+o.proofTree;return l+="\\RightLabel{\\scriptsize(list)}\n",l+="\\BinaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+s.toTexString()+" $}\n",new p(a.eqs.concat(o.eqs,new r.TypeEquation(s,e)),l)}toTexString(t){let e=this.head.toTexString(!1)+"::"+this.tail.toTexString(!0);return t||(e="("+e+")"),e}getRedexes(t,e,n){if(t)return[];throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className)}getLeftMostRedex(t,e,n){if(t)return null;throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className)}extractMacros(){return new q(this.head.extractMacros(),this.tail.extractMacros())}}class R extends g{constructor(t,e,n){super("If"),this.state=t,this.ifTrue=e,this.ifFalse=n}getFV(){return this.freevars=x.union(this.state.getFV(),this.ifTrue.getFV(),this.ifFalse.getFV())}toString(t){let e="[if]"+this.state.toString(!0)+"[then]"+this.ifTrue.toString(!0)+"[else]"+this.ifFalse.toString(!0);return t||(e="("+e+")"),e}substitute(t,e){return new R(this.state.substitute(t,e),this.ifTrue.substitute(t,e),this.ifFalse.substitute(t,e))}equals(t){return t instanceof R&&t.state.equals(this.state)&&t.ifTrue.equals(this.ifTrue)&&t.ifFalse.equals(this.ifFalse)}equalsAlpha(t){return t instanceof R&&t.state.equalsAlpha(this.state)&&t.ifTrue.equalsAlpha(this.ifTrue)&&t.ifFalse.equalsAlpha(this.ifFalse)}getEquations(t,e,n){let i=this.state.getEquations(t,new r.TypeBool,!0),s=this.ifTrue.getEquations(t,e,!0),a=this.ifFalse.getEquations(t,e,!0),o=i.proofTree+s.proofTree+a.proofTree;return o+="\\RightLabel{\\scriptsize(if)}\n",o+="\\TrinaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p(i.eqs.concat(s.eqs,a.eqs),o)}toTexString(t){let e="{\\bf if}~"+this.state.toTexString(!0)+"~{\\bf then}~"+this.ifTrue.toTexString(!0)+"~{\\bf else}~"+this.ifFalse.toTexString(!0);return t||(e="("+e+")"),e}getRedexes(t,e,n){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);return this.state instanceof T?this.state.value?[new f(this,this.ifTrue,"if2")]:[new f(this,this.ifFalse,"if3")]:h.makeNext(this.state.getRedexes(!0,!1,!1),"([if]","[then]"+this.ifTrue.toString(!0)+"[else]"+this.ifFalse.toString(!0)+")","({\\bf if}~","~{\\bf then}~"+this.ifTrue.toTexString(!0)+"~{\\bf else}~"+this.ifFalse.toTexString(!0)+")",t=>new R(t,this.ifTrue,this.ifFalse))}getLeftMostRedex(t,e,n){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);if(this.state instanceof T)return this.state.value?new f(this,this.ifTrue,"if2"):new f(this,this.ifFalse,"if3");{let t=this.state.getLeftMostRedex(!0,!1,!1);return null===t?null:(t.next=new R(t.next,this.ifTrue,this.ifFalse),t.addLeft("([if]"),t.addRight("[then]"+this.ifTrue.toString(!0)+"[else]"+this.ifFalse.toString(!0)+")"),t.addTexLeft("({\\bf if}~"),t.addTexRight("~{\\bf then}~"+this.ifTrue.toTexString(!0)+"~{\\bf else}~"+this.ifFalse.toTexString(!0)+")"),t)}}static parse(t,e){let n=[],r=0,s=0,o=0;for(;;){if(0==t.length)throw new i.LambdaParseError("Illegal If statement");let e=t.shift();switch(e.name){case"if":r++;break;case"then":s++;break;case"else":o++}if(r===o&&s===r+1)break;n.push(e)}let l=a(n,e),h=[];for(r=0,s=0,o=0;;){if(0==t.length)throw new i.LambdaParseError("Illegal If statement");let e=t.shift();switch(e.name){case"if":r++;break;case"then":s++;break;case"else":o++}if(r===s&&o===r+1)break;h.push(e)}let u=a(h,e),c=a(t,e);return new R(l,u,c)}extractMacros(){return new R(this.state.extractMacros(),this.ifTrue.extractMacros(),this.ifFalse.extractMacros())}}class M extends g{constructor(t,e,n){super("Let"),this.boundvar=t,this.left=e,this.right=n}getFV(){if(void 0!==this.freevars)return this.freevars;let t=[];for(let e of this.right.getFV())e.equals(this.boundvar)||t.push(e);return this.freevars=x.union(t,this.left.getFV())}toString(t){let e="[let]"+this.boundvar.toString(!0)+"="+this.left.toString(!0)+"[in]"+this.right.toString(!0);return t||(e="("+e+")"),e}substitute(t,e){let n=this.left.substitute(t,e);if(this.boundvar.equals(t))return new M(this.boundvar,n,this.right);if(x.contains(e.getFV(),this.boundvar)){let r=x.union(this.right.getFV(),e.getFV()),i=x.getNew(r);return i.equals(t)?new M(i,n,this.right.substitute(this.boundvar,i)):new M(i,n,this.right.substitute(this.boundvar,i).substitute(t,e))}return new M(this.boundvar,n,this.right.substitute(t,e))}equals(t){return t instanceof M&&t.boundvar.equals(this.boundvar)&&t.left.equals(this.left)&&t.right.equals(this.right)}equalsAlpha(t){if(!(t instanceof M))return!1;if(this.equals(t))return!0;let e=this.boundvar,n=this.right,r=t.boundvar,i=t.right;return!x.contains(n.getFV(),r)&&i.equalsAlpha(n.substitute(e,r))}getEquations(t,e,n){let i=r.TypeVariable.getNew();this.boundvar.type=i;let s=this.left.getEquations(t,i,!0),a=this.right.getEquations(t.concat(this.boundvar),e,!0),o=s.proofTree+a.proofTree;return o+="\\RightLabel{\\scriptsize(let)}\n",o+="\\TrinaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p(s.eqs.concat(a.eqs),o)}toTexString(t){let e="{\\bf let}~"+this.boundvar.toTexString(!1)+" = "+this.left.toTexString(!0)+"~{\\bf in}~"+this.right.toTexString(!0);return t||(e="("+e+")"),e}getRedexes(t,e,n){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);return[new f(this,this.right.substitute(this.boundvar,this.left),"let")]}getLeftMostRedex(t,e,n){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);return new f(this,this.right.substitute(this.boundvar,this.left),"let")}static parse(t,e){let n=t.shift();if(null===n.name.match(/^[A-Za-z]$/))throw new i.LambdaParseError("Unexpected token: '"+n+"'");let r=new x(n.name);if("="!==t.shift().name)throw new i.LambdaParseError("'=' is expected");let s=[],o=1;for(;;){if(0==t.length)throw new i.LambdaParseError("Illegal Let statement");let e=t.shift();if("let"===e.name?o++:"in"===e.name&&o--,0==o)break;s.push(e)}let l=a(s,e),h=a(t,e);return new M(r,l,h)}extractMacros(){return new M(this.boundvar,this.left.extractMacros(),this.right.extractMacros())}}class N extends g{constructor(t,e,n,r,i){super("Case"),this.state=t,this.ifNil=e,this.head=n,this.tail=r,this.ifElse=i}getFV(){return void 0!==this.freevars?this.freevars:x.union(this.state.getFV(),this.ifNil.getFV(),x.dif(this.ifElse.getFV(),[this.head,this.tail]))}toString(t){let e="[case]"+this.state.toString(!0)+"[of][nil]->"+this.ifNil.toString(!0)+"|"+this.head.toString(!0)+"::"+this.tail.toString(!0)+"->"+this.ifElse.toString(!0);return t||(e="("+e+")"),e}substitute(t,e){let n=this.state.substitute(t,e),r=this.ifNil.substitute(t,e);if(this.head.equals(t)||this.tail.equals(t))return new N(n,r,this.head,this.tail,this.ifElse);if(x.contains(e.getFV(),this.head)||x.contains(e.getFV(),this.tail)){let i=this.head,s=this.tail,a=this.ifElse;if(x.contains(e.getFV(),i)){let n=x.union(this.ifElse.getFV(),e.getFV()),r=x.getNew(n);a=r.equals(t)?a.substitute(i,r):a.substitute(i,r).substitute(t,e),i=r}if(x.contains(e.getFV(),s)){let n=x.union(this.ifElse.getFV(),e.getFV()),r=x.getNew(n);a=r.equals(t)?a.substitute(s,r):a.substitute(s,r).substitute(t,e),s=r}return new N(n,r,i,s,a)}return new N(n,r,this.head,this.tail,this.ifElse.substitute(t,e))}equals(t){return t instanceof N&&t.state.equals(this.state)&&t.ifNil.equals(this.ifNil)&&t.head.equals(this.head)&&t.tail.equals(this.tail)&&t.ifElse.equals(this.ifElse)}equalsAlpha(t){return t instanceof N&&t.state.equalsAlpha(this.state)&&t.ifNil.equalsAlpha(this.ifNil)&&t.head.equalsAlpha(this.head)&&t.tail.equalsAlpha(this.tail)&&t.ifElse.equalsAlpha(this.ifElse)}getEquations(t,e,n){let i=r.TypeVariable.getNew(),s=new r.TypeList(i);this.head.type=i,this.tail.type=s;let a=this.state.getEquations(t,s,!0),o=this.ifNil.getEquations(t,e,!0),l=this.ifElse.getEquations(t.concat(this.head,this.tail),e,!0),h=a.proofTree+o.proofTree+l.proofTree;return h+="\\RightLabel{\\scriptsize(case)}\n",h+="\\TrinaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p(a.eqs.concat(o.eqs,l.eqs),h)}toTexString(t){let e="{\\bf case} "+this.state.toTexString(!0)+" {\\bf of} {\\rm nil} \\Rightarrow "+this.ifNil.toTexString(!0)+" | "+this.head.toTexString(!0)+"::"+this.tail.toTexString(!0)+" \\Rightarrow "+this.ifElse.toTexString(!0);return t||(e="("+e+")"),e}getRedexes(t,e){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);return this.state instanceof v?[new f(this,this.ifNil,"case2")]:this.state instanceof q?[new f(this,this.ifElse.substitute(this.head,this.state.head).substitute(this.tail,this.state.tail),"case3")]:h.makeNext(this.state.getRedexes(!0,!1,!0),"([case]","[of][nil]->"+this.ifNil+" | "+this.head+"::"+this.tail+"->"+this.ifElse+")","({\\bf case} "," {\\bf of} {\\rm nil} \\Rightarrow "+this.ifNil.toTexString(!0)+" | "+this.head.toTexString(!0)+"::"+this.tail.toTexString(!0)+" \\Rightarrow "+this.ifElse.toTexString(!0)+")",t=>new N(t,this.ifNil,this.head,this.tail,this.ifElse))}getLeftMostRedex(t,e){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);if(this.state instanceof v)return new f(this,this.ifNil,"case2");if(this.state instanceof q)return new f(this,this.ifElse.substitute(this.head,this.state.head).substitute(this.tail,this.state.tail),"case3");{let t=this.state.getLeftMostRedex(!0,!1,!0);return null===t?null:(t.next=new N(t.next,this.ifNil,this.head,this.tail,this.ifElse),t.addLeft("([case]"),t.addRight("[of][nil]->"+this.ifNil+" | "+this.head+"::"+this.tail+"->"+this.ifElse+")"),t.addTexLeft("({\\bf case} "),t.addTexRight(" {\\bf of} {\\rm nil} \\Rightarrow "+this.ifNil.toTexString(!0)+" | "+this.head.toTexString(!0)+"::"+this.tail.toTexString(!0)+" \\Rightarrow "+this.ifElse.toTexString(!0)+")"),t)}}static parse(t,e){let n=[],r=1;for(;;){if(0==t.length)throw new i.LambdaParseError("Illegal Case statement");let e=t.shift();if("case"===e.name?r++:"of"===e.name&&r--,0==r)break;n.push(e)}let s=a(n,e),o=t.shift();if("nil"!==o.name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");if("-"!==(o=t.shift()).name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");if(">"!==(o=t.shift()).name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");let l=[];for(r=1;;){if(0==t.length)throw new i.LambdaParseError("Too many [case]");let e=t.shift();if("case"===e.name?r++:"|"===e.name&&r--,0==r)break;l.push(e)}let h=a(l,e),u=new x(t.shift().name);if(null===u.name.match(/^[A-Za-z]$/))throw new i.LambdaParseError("Unexpected token: '"+u.name+"'");if(":"!==(o=t.shift()).name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");if(":"!==(o=t.shift()).name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");let c=new x(t.shift().name);if(null===c.name.match(/^[A-Za-z]$/))throw new i.LambdaParseError("Unexpected token: '"+c.name+"'");if("-"!==(o=t.shift()).name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");if(">"!==(o=t.shift()).name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");let d=a(t,e);return new N(s,h,u,c,d)}extractMacros(){return new N(this.state.extractMacros(),this.ifNil.extractMacros(),this.head,this.tail,this.ifElse.extractMacros())}}class F extends g{constructor(t,e){super("Fix"),this.boundvar=t,this.expr=e}getFV(){if(void 0!==this.freevars)return this.freevars;let t=[];for(let e of this.expr.getFV())e.equals(this.boundvar)||t.push(e);return this.freevars=t}toString(t){let e="[fix]"+this.boundvar.toString(!1)+"."+this.expr.toString(!0);return t||(e="("+e+")"),e}substitute(t,e){if(this.boundvar.equals(t))return this;if(x.contains(e.getFV(),this.boundvar)){let n=x.union(this.expr.getFV(),e.getFV()),r=x.getNew(n);return new F(r,this.expr.substitute(this.boundvar,r)).substitute(t,e)}return new F(this.boundvar,this.expr.substitute(t,e))}equals(t){return t instanceof F&&t.boundvar.equals(this.boundvar)&&t.expr.equals(this.expr)}equalsAlpha(t){if(!(t instanceof F))return!1;if(this.equals(t))return!0;let e=this.boundvar,n=this.expr,r=t.boundvar,i=t.expr;return x.contains(n.getFV(),r)?i.equalsAlpha(n):i.equalsAlpha(n.substitute(e,r))}getEquations(t,e,n){this.boundvar.type=e;let r=this.expr.getEquations(t.concat(this.boundvar),e,!0),i=r.proofTree;return i+="\\RightLabel{\\scriptsize(fix)}\n",i+="\\UnaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p(r.eqs,i)}toTexString(t){let e="{\\bf fix}~"+this.boundvar.toTexString(!0)+"."+this.expr.toTexString(!0);return t||(e="("+e+")"),e}getRedexes(t,e,n){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);return[new f(this,this.expr.substitute(this.boundvar,new F(new x(this.boundvar.name),this.expr)),"fix")]}getLeftMostRedex(t,e,n){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);return new f(this,this.expr.substitute(this.boundvar,new F(new x(this.boundvar.name),this.expr)),"fix")}static parse(t,e){let n=t.shift();if(null===n.name.match(/^[A-Za-z]$/))throw new i.LambdaParseError("Unexpected token: '"+n+"'");let r=new x(n.name);if("."!==t.shift().name)throw new i.LambdaParseError("'.' is expected");let s=a(t,e);return new F(r,s)}extractMacros(){return new F(this.boundvar,this.expr.extractMacros())}}},function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});const r=n(1),i=n(0);class s{constructor(t,e){this.children=[],this.label=t,this.info=e,this.id=e.nextId++,e.nodes.push(this)}toString(){return this.label}equalsShape(t){return this.info.nodes.length===t.info.nodes.length&&this.info.edges.length===t.info.edges.length&&this.info.allowMultipleEdges===t.info.allowMultipleEdges&&function t(e,n,r){if(e.children.length!==n.children.length)return!1;r.push({n1:e,n2:n});let i=[].concat(n.children);let s=!0;for(let n of e.children){let e=!1;for(let s=0;s<i.length;s++){let a=i[s],o=void 0;for(let t of r)if(t.n1.id===n.id||t.n2.id===a.id){o=t;break}if(void 0===o&&t(n,a,r)||void 0!==o&&o.n1.id===n.id&&o.n2.id===a.id){i.splice(s,1),e=!0;break}}if(!e){s=!1;break}}for(let t=0;t<r.length;t++)if(r[t].n1.id===e.id)return r.splice(t,1),s;throw new Error("Unexpected Error")}(this,t,[])}static parse(t){let e=t.split(";"),n=new a([],[],null,null,0,!1),i={};for(let t of e){let e=t.split("->"),r=o(e[0]);for(let t=1;t<e.length;t++){let i=o(e[t]);for(let t of r)for(let e of i)n.hasEdge({from:t,to:e})&&(n.allowMultipleEdges=!0),n.edges.push({from:t,to:e}),t.children.push(e);r=i}}if(void 0===i.root)throw new r.GraphParseError("there must be a 'root' node.");return i.root;function o(t){let e=t.split(","),a=[];for(let t of e){if(""===(t=t.trim()))continue;if(!/^[a-z][a-z0-9]*$/.test(t))throw new r.GraphParseError("the name '"+t+"' cannot be used as node name, must be [a-z][a-z0-9]*");let e=i[t];void 0===e&&(e=i[t]=new s(t,n)),a.push(e)}return a}}static search(t,e){let r,s="graph_closure.csv";try{let t=new XMLHttpRequest;t.open("GET",s,!1),t.send(null),r=t.responseText}catch(t){0;let e=n(6);try{e.statSync(s)}catch(t){return void console.error("File Not Found: "+s)}r=e.readFileSync(s,"utf8")}let a=r.split("\n");for(let n of a){let r=n.split(",");if(parseInt(r[2])===t.info.nodes.length&&parseInt(r[3])===t.info.edges.length){let n=new i.LambdaFriends(r[0],!1,!1,e);for(;null!==n.deepen(););if(n.root.equalsShape(t))return n}}return null}}e.GraphNode=s;class a{constructor(t,e,n,r,i,s){this.nodes=t,this.edges=e,this.typed=n,this.etaAllowed=r,this.nextId=i,this.allowMultipleEdges=s}hasEdge(t){for(let e of this.edges)if(e.from.id===t.from.id&&e.to.id===t.to.id)return!0;return!1}}class o extends s{constructor(t,e,n){super((t=t.extractMacros()).toString(!0),n),this.expr=t,this.parent=e,this.depth=null===e?0:e.depth+1,this.nextrs=this.expr.getRedexes(n.typed,n.etaAllowed,!0),this.isNormalForm=0===this.nextrs.length}static makeRoot(t,e,n,r){return new o(t,null,new a([],[],e,n,0,r))}visit(){let t={nodes:[],edges:[]};for(let e of this.nextrs){let n=this.find(e.next);if(null===n){let n=new o(e.next,this,this.info);this.children.push(n),t.nodes.push(n),!this.info.allowMultipleEdges&&this.info.hasEdge({from:this,to:n})||(t.edges.push({from:this,to:n}),this.info.edges.push({from:this,to:n}))}else this.children.push(n),!this.info.allowMultipleEdges&&this.info.hasEdge({from:this,to:n})||(t.edges.push({from:this,to:n}),this.info.edges.push({from:this,to:n}))}return t}find(t){for(let e of this.info.nodes)if(e.expr.equalsAlpha(t))return e;return null}}e.ReductionNode=o},function(t,e){t.exports=require("fs")}]);
+!function(t){var e={};function n(r){if(e[r])return e[r].exports;var i=e[r]={i:r,l:!1,exports:{}};return t[r].call(i.exports,i,i.exports,n),i.l=!0,i.exports}n.m=t,n.c=e,n.d=function(t,e,r){n.o(t,e)||Object.defineProperty(t,e,{enumerable:!0,get:r})},n.r=function(t){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(t,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(t,"__esModule",{value:!0})},n.t=function(t,e){if(1&e&&(t=n(t)),8&e)return t;if(4&e&&"object"==typeof t&&t&&t.__esModule)return t;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:t}),2&e&&"string"!=typeof t)for(var i in t)n.d(r,i,function(e){return t[e]}.bind(null,i));return r},n.n=function(t){var e=t&&t.__esModule?function(){return t.default}:function(){return t};return n.d(e,"a",e),e},n.o=function(t,e){return Object.prototype.hasOwnProperty.call(t,e)},n.p="",n(n.s=3)}([function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});const r=n(4),i=n(2),s=n(5);class a{constructor(t,e,n,i){let a=t.split("#")[0].trim(),o=[];for(;;){let t=a.match(/^[^[]+?\s*=\s*/);if(null===t)break;let e=t[0];(t=a.split(e)).shift(),a=t.join(e),o.push(e.split(/\s*=\s*$/)[0])}this.expr=r.makeAST(a,e),this.original=this.expr,this.type=this.getType(e);for(let t of o)r.Macro.add(t,this,e);this.typed=e,this.etaAllowed=n,this.allowMultipleEdges=i,this.processTex="\\begin{eqnarray*}\n&& ",this.curStep=0,this.root=s.ReductionNode.makeRoot(this.expr,this.typed,this.etaAllowed,this.allowMultipleEdges),this.curNodes=[this.root],this.nextRedexes=void 0}getRedexes(){return this.nextRedexes?this.nextRedexes:this.nextRedexes=this.expr.getRedexes(this.typed,this.etaAllowed,!0).sort(r.Redex.compare)}getLeftMostRedex(){return this.typed?this.getRedexes()[0]||null:this.nextLeftMostRedex?this.nextLeftMostRedex:this.nextLeftMostRedex=this.expr.getLeftMostRedex(this.typed,this.etaAllowed,!0)}reduction(t){if(void 0===t&&!(t=this.getLeftMostRedex()))return null;let e;if(this.expr=t.next,this.nextRedexes=void 0,this.nextLeftMostRedex=void 0,this.processTex+=t.toTexString(),"macro"===t.type?(this.processTex+=" \\\\\n&\\equiv& ",e="-: (macro) = "+this.expr.toString(!0)):(this.processTex+=" \\\\\n&\\longrightarrow_{"+t.getTexRule()+"}& ",e=++this.curStep+": ("+t.rule+") --\x3e "+this.expr.toString(!0)),!this.hasNext()){e+="    (normal form)\n";let t=this.parseChurchNum();null!==t&&(e+="  = "+t+" (as nat)\n");let n=this.parseChurchBool();null!==n&&(e+="  = "+n+" (as bool)\n"),e=e.slice(0,-1)}return e}deepen(t){if(0===this.curNodes.length)return null;let e=this.curNodes.shift();if(void 0!==t&&e.depth>=t)return this.curNodes.push(e),null;let n=e.visit();for(let t of n.nodes)this.curNodes.push(t);return n}hasNext(){return null!==this.getLeftMostRedex()}hasNodes(){return this.curNodes.length>0}getProofTree(){return"\\begin{prooftree}\n"+this.proofTree+"\\end{prooftree}"}getProcessTex(){return this.processTex+this.expr.toTexString(!0)+(this.hasNext()?"":"\\not\\longrightarrow")+"\n\\end{eqnarray*}"}getType(t){if(!t)return new i.TypeUntyped;i.TypeVariable.maxId=void 0;let e=i.TypeVariable.getNew(),n=this.expr.getEquations([],e,!0),r=n.eqs;this.proofTree=n.proofTree;let s=i.TypeEquation.get(e,i.TypeEquation.solve(r)),a=s.getVariables(),o=[];for(let t of a)i.TypeVariable.contains(o,t)||o.push(t);let l=0;for(let t of o)s.replace(t,i.TypeVariable.getAlphabet(l)),l++;return s}static parseMacroDef(t,e){let n=t.split("#")[0].trim(),i=[];for(;;){let t=n.match(/^[^[]+?\s*=\s*/);if(null===t)break;let e=t[0];(t=n.split(e)).shift(),n=t.join(e),i.push(e.split(/\s*=\s*$/)[0])}if(0===i.length)return null;let s=new a(n,e,void 0,!1);for(let t of i)r.Macro.add(t,s,e);return{names:i,expr:s.expr.toString(!0),type:s.type.toString()}}static fileInput(t,e){let n=t.split("\n"),r=[],i=[];for(let t of n)try{let n=a.parseMacroDef(t,e);null!==n&&i.push(n)}catch(t){r.push(t.toString())}return{defs:i,errs:r}}static getMacroList(t){let e="",n=r.Macro.getMap(t);for(let t in n){let r=n[t];e+="<"+r.name+"> is defined as "+r.expr.toString(!0)+" : "+r.type+"\n"}return e}static getMacroListAsObject(t){return r.Macro.getMap(t)}static clearMacro(t){return r.Macro.clear(t)}static graph2LF(t,e){return s.GraphNode.search(s.GraphNode.parse(t),e)}static lmntal2LF(t){return new a(r.parseLMNtal(t).toString(!0),!1,!1,!1)}toLMNtal(){return a.nextLinkID=0,this.typed?null:this.original.toLMNtal()}toString(){let t=this.expr.toString(!0)+" : "+this.type;if(!this.hasNext()){t+="    (normal form)\n";let e=this.parseChurchNum();null!==e&&(t+="  = "+e+" (as nat)\n");let n=this.parseChurchBool();null!==n&&(t+="  = "+n+" (as bool)\n"),t=t.slice(0,t.length-1)}return t}getOriginalString(){return this.original+" : "+this.type}parseChurchNum(){return this.expr.parseChurchNum()}parseChurchBool(){return this.expr.parseChurchBool()}static getNewLink(){return"R"+a.nextLinkID++}}e.LambdaFriends=a},function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});class r{constructor(t,e){void 0===e?(this.name="LambdaFriendsError",this.message=t):(this.name=t,this.message=e)}toString(){return this.name+": "+this.message}}e.LambdaFriendsError=r;e.LambdaParseError=class extends r{constructor(t){super("LambdaParseError",t)}};e.SubstitutionError=class extends r{constructor(t){super("SubstitutionError",t)}};e.ReductionError=class extends r{constructor(t){super("ReductionError",t)}};e.MacroError=class extends r{constructor(t){super("MacroError",t)}};e.TypeError=class extends r{constructor(t){super("TypeError",t)}};e.TexError=class extends r{constructor(t){super("TexError",t)}};e.GraphParseError=class extends r{constructor(t){super("GraphParseError",t)}}},function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});const r=n(1);class i{constructor(t){this.className=t}}e.Type=i;class s{constructor(t,e){this.left=t,this.right=e}transform(t,e){if(this.left instanceof a&&this.right instanceof a)return this.left.match(this.right);if(this.left.equals(this.right))return[];if(!(this.left instanceof c))return[new s(this.right,this.left)];if(this.right.contains(this.left))throw new r.TypeError("Illegal type ("+this.right+" contains "+this.left+". Self-application?)");for(var n of t)n.replace(this.left,this.right);for(var n of e)n.replace(this.left,this.right);return[this]}static isEqual(t,e){if(t.length!==e.length)return!1;for(var n of t)if(!s.contains(n,e))return!1;return!0}static contains(t,e){for(var n of e)if(t.equals(n))return!0;return!1}equals(t){return t.left.equals(this.left)&&t.right.equals(this.right)}static get(t,e){for(var n of e)if(n.left.equals(t))return n.right;throw new r.TypeError("Undefined TypeVariable: "+t)}toString(){return this.left+" = "+this.right}replace(t,e){this.left.equals(t)?this.left=e:this.left.replace(t,e),this.right.equals(t)?this.right=e:this.right.replace(t,e)}static solve(t){for(;;){for(var e=[].concat(t),n=[];t.length>0;){var r=t.shift().transform(t,n);n=n.concat(r)}if(t=[].concat(n),s.isEqual(e,n))break}return t}}e.TypeEquation=s;class a extends i{}e.TypeConstructor=a;class o extends a{constructor(){super("TypeInt")}toString(){return"int"}toTexString(){return"{\\rm int}"}equals(t){return t instanceof o}match(t){if(t instanceof o)return[];throw new r.TypeError(this+" and "+t+" are not compatible.")}contains(t){return!1}replace(t,e){}getVariables(){return[]}}e.TypeInt=o;class l extends a{constructor(){super("TypeBool")}toString(){return"bool"}toTexString(){return"{\\rm bool}"}equals(t){return t instanceof l}match(t){if(t instanceof l)return[];throw new r.TypeError(this+" and "+t+" are not compatible.")}contains(t){return!1}replace(t,e){}getVariables(){return[]}}e.TypeBool=l;class h extends a{constructor(t){super("TypeList"),this.content=t}toString(){return"list("+this.content+")"}toTexString(){return"{\\rm list}("+this.content.toTexString()+")"}equals(t){return t instanceof h&&this.content.equals(t.content)}match(t){if(t instanceof h)return[new s(this.content,t.content)];throw new r.TypeError(this+" and "+t+" are not compatible.")}contains(t){return this.content.contains(t)}replace(t,e){this.content.equals(t)?this.content=e:this.content.replace(t,e)}getVariables(){return this.content.getVariables()}}e.TypeList=h;class u extends a{constructor(t,e){super("TypeFunc"),this.left=t,this.right=e}toString(){return(this.left instanceof u?"("+this.left+")":this.left.toString())+" -> "+this.right}toTexString(){return(this.left instanceof u?"("+this.left.toTexString()+")":this.left.toTexString())+" \\rightarrow "+this.right.toTexString()}equals(t){return t instanceof u&&(t.left.equals(this.left)&&t.right.equals(this.right))}match(t){if(t instanceof u)return[new s(this.left,t.left),new s(this.right,t.right)];throw new r.TypeError(this+" and "+t+" are not compatible.")}contains(t){return this.left.contains(t)||this.right.contains(t)}replace(t,e){this.left.equals(t)?this.left=e:this.left.replace(t,e),this.right.equals(t)?this.right=e:this.right.replace(t,e)}getVariables(){return this.left.getVariables().concat(this.right.getVariables())}}e.TypeFunc=u;class c extends i{constructor(t){super("TypeVariable"),this.id=t}toString(){return this.id<0?"'"+c.alphabet[-this.id-1]:"'t"+this.id}toTexString(){return this.id<0?c.texAlphabet[-this.id-1]:"\\tau_{"+this.id+"}"}equals(t){return t instanceof c&&this.id===t.id}static getNew(){return void 0===c.maxId?(c.maxId=0,new c(0)):(c.maxId++,new c(c.maxId))}contains(t){return this.equals(t)}replace(t,e){}getVariables(){return[this]}static getAlphabet(t){return new c(-t-1)}static contains(t,e){for(var n of t)if(n.equals(e))return!0;return!1}}c.alphabet="abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ".split(""),c.texAlphabet=["\\alpha","\\beta","\\gamma","\\delta","\\varepsilon","\\zeta","\\eta","\\theta","\\iota","\\kappa","\\mu","\\nu","\\xi","\\pi","\\rho","\\sigma","\\upsilon","\\phi","\\chi","\\psi","\\omega","\\Gamma","\\Delta","\\Theta","\\Xi","\\Pi","\\Sigma","\\Phi","\\Psi","\\Omega"].concat(c.alphabet),e.TypeVariable=c;class d extends i{constructor(){super("TypeUntyped")}toString(){return"Untyped"}toTexString(){return"{\\rm Untyped}"}equals(t){return t instanceof d}contains(t){return!1}replace(t,e){}getVariables(){return[]}}e.TypeUntyped=d},function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});const r=n(0);let i=cytoscape({container:document.getElementById("graph"),boxSelectionEnabled:!1,autounselectify:!0,style:[{selector:"node",style:{label:"data(label)","background-color":"#11479e"}},{selector:"edge",style:{"target-arrow-shape":"triangle","curve-style":"bezier","target-arrow-color":"#9dbaea",width:3,"line-color":"#9dbaea"}},{selector:".goal",style:{"background-color":"#b3424a"}}]});MicroModal.init({disableScroll:!1,awaitCloseAnimation:!0});let s=void 0,a=!0,o=!1,l=!1,h=void 0,u=document.getElementById("input"),c=document.getElementById("output"),d=document.getElementById("untyped"),f=document.getElementById("typed"),p=document.getElementById("etaEnable"),g=document.getElementById("etaDisable"),m=document.getElementById("multiEdgeEnable"),x=document.getElementById("multiEdgeDisable"),w=document.getElementById("fileInput"),b=new FileReader,T=document.getElementById("clearMacroBtn"),y=document.getElementById("translate"),v=document.getElementById("tabA"),S=document.getElementById("outputBtns"),E=document.getElementById("stepInput"),L=document.getElementById("startGraph"),q=document.getElementById("stopGraph"),R=document.getElementById("imgGraph"),M=document.getElementById("maxDepth"),N=document.getElementById("tabAbtn"),F=document.getElementById("tabBbtn"),k=document.getElementById("tabCbtn"),I=document.getElementById("tabDbtn"),B=document.getElementById("lmnInput"),A=document.getElementById("lmnSubmit"),V=document.getElementById("lmnOutput"),C=document.getElementById("graphInput"),P=document.getElementById("graphSubmit"),U=document.getElementById("graphOutput");w.addEventListener("change",function(t){let e=t.target.files[0];if("text/plain"!==e.type)return alert("プレーンテキストを選択してください"),void(w.value="");b.readAsText(e)}),b.addEventListener("load",function(){let t=r.LambdaFriends.fileInput(b.result,a);tt();let e=document.getElementById("fileInputLog");if(e.textContent=null,t.defs.length>0){let n=document.createElement("div"),r=document.createElement("p");r.innerText="Finally, "+t.defs.length+" macros are successfully added.";let i=document.createElement("ul");i.className="code";for(let e of t.defs){let t=document.createElement("li"),n=[].concat(e.names),r="<"+n.shift()+">";for(;n.length>0;){r+=" = <"+n.shift()+">"}t.innerText=r+" is defined as "+e.expr+" : "+e.type,i.appendChild(t)}n.appendChild(r),n.appendChild(i),e.appendChild(n)}if(t.errs.length>0){let n=document.createElement("div"),r=document.createElement("p");r.innerText="Unfortunately, "+t.errs.length+" macros are rejected due to some errors.";let i=document.createElement("ul");i.className="code";for(let e of t.errs){let t=document.createElement("li");t.innerText=e,i.appendChild(t)}n.appendChild(r),n.appendChild(i),e.appendChild(n)}MicroModal.show("modal-1",{debugMode:!0,disableScroll:!0,awaitCloseAnimation:!0})}),d.onclick=function(){d.className="btn btn-primary",f.className="btn btn-default",a=!1,p.disabled=!1,g.disabled=!1,m.disabled=!1,x.disabled=!1,tt()},f.onclick=function(){f.className="btn btn-primary",d.className="btn btn-default",a=!0,p.disabled=!0,g.disabled=!0,m.disabled=!0,x.disabled=!0,tt()},p.onclick=function(){p.className="btn btn-primary",g.className="btn btn-default",o=!0},g.onclick=function(){g.className="btn btn-primary",p.className="btn btn-default",o=!1},m.onclick=function(){m.className="btn btn-primary",x.className="btn btn-default",l=!0},x.onclick=function(){x.className="btn btn-primary",m.className="btn btn-default",l=!1},T.onclick=function(){r.LambdaFriends.clearMacro(a),tt()},E.addEventListener("change",function(){let t=parseInt(E.value);s=isNaN(t)?void 0:t});let $=[],j=0,z=[""],O=function(){let t=u.value;if(""!==t||void 0===h){$.unshift(t),j=0,(z=[].concat($)).unshift(""),t=(t=t.split("#")[0]).trim();try{let e=r.LambdaFriends.parseMacroDef(t,a);if(null===e)h=new r.LambdaFriends(t,a,o,l),H=0,i.remove("*"),i.add({group:"nodes",data:{id:""+h.root.id,label:h.root.toString()},classes:h.root.isNormalForm?"goal":""}),ot(),W(h.toString()),a&&st(),at();else{let t=[].concat(e.names),n="<"+t.shift()+">";for(;t.length>0;){n+=" = <"+t.shift()+">"}W(n+=" is defined as "+e.expr+" : "+e.type),S.textContent=null}nt()}catch(t){W(t.toString()),console.log(t),S.textContent=null}tt(),u.value=""}else X?Z():st()};document.getElementById("submit").onclick=O,document.getElementById("input").onkeydown=function(t){13===t.keyCode?(O(),t.preventDefault()):38===t.keyCode?(j<z.length-1&&(z[j]=u.value,j++,u.value=z[j]),t.preventDefault()):40===t.keyCode&&(j>0&&(z[j]=u.value,j--,u.value=z[j]),t.preventDefault())};let _,D=!1,H=0,G=!1;function Z(){if(G)return;if(ot(),void 0===h)return;D=!1,G=!0,H+=_||10;let t=()=>setTimeout(()=>{if(D)return ot(),void(G=!1);let e=h.deepen(H);if(null===e)return ot(),void(G=!1);let n=[];for(let t of e.nodes)n.push({group:"nodes",data:{id:""+t.id,label:""+t.toString()},classes:t.isNormalForm?"goal":""});for(let t of e.edges)n.push({group:"edges",data:{source:t.from.id.toString(),target:t.to.id.toString()}});ot(),i.add(n),t(),ot()},1);t()}L.onclick=Z,q.onclick=function(){D=!0},R.onclick=function(){window.open(i.png())},M.addEventListener("change",function(){let t=parseInt(M.value);_=isNaN(t)?void 0:t});let X=!1;function K(){let t=B.value;try{let e=r.LambdaFriends.lmntal2LF(t);V.innerText="Found: "+e.expr.toString(!0)}catch(t){V.innerText=t.toString()}}function J(){let t=C.value;try{let e=r.LambdaFriends.graph2LF(t,l);U.innerText=null===e?"Not Found":"Found: "+e.expr.toString(!0)}catch(t){U.innerText=t.toString()}}N.addEventListener("click",()=>{X=!1}),F.addEventListener("click",()=>{X=!1}),k.addEventListener("click",()=>{X=!1}),I.addEventListener("click",()=>{setTimeout(ot,10),X=!0}),B.onkeydown=function(t){13===t.keyCode&&(K(),t.preventDefault())},A.onclick=K,C.onkeydown=function(t){13===t.keyCode&&(J(),t.preventDefault())},P.onclick=J;let Q="";function W(t){Q=t+"\n",c.innerHTML=et(Q)}function Y(t){Q+=t+"\n",c.innerHTML=et(Q)}function tt(){let t=document.getElementById("macroList");t.innerHTML="";let e=r.LambdaFriends.getMacroListAsObject(a);for(let n in e){let r=e[n];t.innerHTML+="<tr><th>"+et(r.name)+"</th><td>"+et(r.expr.toString(!0))+"</td><td>"+et(r.type.toString())+"</td></tr>"}}function et(t){return t.replace(/[&'`"<> \n]/g,function(t){return{"&":"&amp;","'":"&#x27;","`":"&#x60;",'"':"&quot;","<":"&lt;",">":"&gt;"," ":"&nbsp;","\n":"<br>"}[t]})}function nt(){y.textContent=null,void 0!==h&&(y.appendChild(rt("これまでの簡約過程",h.getProcessTex())),a?y.appendChild(rt("型付けの証明木",h.getProofTree())):y.appendChild(rt("LMNtalコード",h.toLMNtal())))}function rt(t,e){let n=document.createElement("p"),r=document.createElement("button"),i=document.createElement("span"),s=document.createElement("div"),a=document.createElement("p");s.classList.add("code"),n.appendChild(r),n.appendChild(i),s.appendChild(a),a.innerText=e,i.innerText=t,r.type="button",r.className="btn btn-default btn-sm",r.innerText="クリップボードにコピー",r.onclick=function(){let t=document.getSelection();t.selectAllChildren(s),document.execCommand("copy"),t.collapse(document.body,0)};let o=document.createElement("div");return o.appendChild(n),o.appendChild(s),o}let it=!1;function st(){if(it)return;it=!0,S.textContent=null;let t=e=>setTimeout(()=>{if(0===e||!h.hasNext())return at(),v.scrollTop=c.offsetHeight-15,void(it=!1);Y(h.reduction()),v.scrollTop=c.offsetHeight-15,nt(),t(e-1)},1);t(void 0===s?100:s)}function at(){if(!h.hasNext())return void(S.textContent=null);let t=document.createElement("button");if(t.type="button",t.className="btn btn-default btn-sm",t.innerText="最左簡約を続ける",t.onclick=st,S.textContent=null,S.appendChild(t),a)return;let e=document.createElement("span");e.innerText="または、以下から簡約基を選ぶ：",S.appendChild(e);let n=document.createElement("div");n.className="list-group",S.appendChild(n);let r=h.getRedexes();for(let t of r){let e=document.createElement("button");e.className="list-group-item code",e.innerHTML=t.toHTMLString(),e.onclick=function(){Y(h.reduction(t)),at(),v.scrollTop=c.offsetHeight-15,nt()},n.appendChild(e)}}function ot(){i.resize(),i.elements().makeLayout({name:"dagre",nodeSpacing:5,animate:!0,randomize:!1,maxSimulationTime:1500}).run()}d.onclick(null),g.onclick(null),x.onclick(null),tt()},function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});const r=n(2),i=n(1),s=n(0);function a(t,e){let n=null;for(;t.length>0;){let r=t.shift();if(r instanceof w||r instanceof v||r instanceof S)n=null===n?r:new L(n,r);else switch(r.name){case"\\":case"¥":case"λ":return null===n?E.parse(t,e):new L(n,E.parse(t,e));case"(":{let r=[],s=1;for(;;){if(0==t.length)throw new i.LambdaParseError("Too many LPAREN '('");let e=t.shift();if("("===e.name?s++:")"===e.name&&s--,0==s)break;r.push(e)}let o=a(r,e);n=null===n?o:new L(n,o);break}default:if(e)switch(r.name){case"if":return R.parse(t,e);case"let":return M.parse(t,e);case"case":return N.parse(t,e);case"fix":return F.parse(t,e);case":":{let r=t.shift();if(":"!==r.name)throw new i.LambdaParseError("Unexpected token: '"+r+"'");return new q(n,a(t,e))}}if(null===r.name.match(/^[A-Za-z]$/))throw new i.LambdaParseError("Unexpected token: '"+r+"'");n=null===n?new x(r.name):new L(n,new x(r.name))}}if(null===n)throw new i.LambdaParseError("No contents in Expression");return n}function o(t,e){return a(function(t,e){let n=t.split(""),r=[];for(;n.length>0;){let t=n.shift().trim();if(""===t);else if("<"===t){let s="";for(;;){if(0==n.length)throw new i.LambdaParseError("Too many LANGLE '<'");if(">"===(t=n.shift()))break;s+=t}r.push(S.get(s,e))}else if(e&&"["===t){let e="";for(;;){if(0==n.length)throw new i.LambdaParseError("Too many LBRACKET '['");if("]"===(t=n.shift()))break;e+=t}let s=null;switch(e){case"nil":s=new v;break;case"false":case"true":s=new T("true"===e);break;case"if":case"then":case"else":case"let":case"in":case"case":case"of":case"fix":s=new m(e);break;default:s=null!==e.match(/^\d+$|^-\d+$/)?new b(parseInt(e)):new y(e)}if(null===s)throw new i.LambdaParseError("Unknown Const: ["+e+"]");r.push(s)}else r.push(new m(t))}return r}(t,e),e)}function l(t){return t.replace(/[&'`"<>]/g,function(t){return{"&":"&amp;","'":"&#x27;","`":"&#x60;",'"':"&quot;","<":"&lt;",">":"&gt;"}[t]})}e.makeAST=o,e.parseLMNtal=function(t){return function t(n,r,s){let a=(n=n.trim()).match(/^.+?(?=\()/);if(null===a){let t=s[n];if(void 0===t)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. Unknown Token: "+n);return new x(t)}let o=a[0].trim(),l=e(n.match(/\(.+$/)[0]);switch(o){case"lambda":{if(2!==l.length)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. lambda(X,A) should have 2 args.");let n=x.getNew(r),a=function t(n){let r=(n=n.trim()).match(/^.+?(?=\()/);if(null===r)return"rm"===n?[]:[n];let s=r[0].trim(),a=e(n.match(/\(.+$/)[0]);if("rm"===s){if(0!==a.length)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. rm() should have 0 args.");return[]}if("cp"===s){if(2!==a.length)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. cp(A,B) should have 2 args.");return t(a[0].trim()).concat(t(a[1].trim()))}throw new i.LambdaParseError("Malformed LMNtal Lambda Term. Unexpected atom name: "+s)}(l[0]);for(let t of a)s[t]=n.name;return new E(n,t(l[1],r.concat(n),Object.assign({},s)))}case"apply":if(2!==l.length)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. apply(A,B) should have 2 args.");return new L(t(l[0],[].concat(r),Object.assign({},s)),t(l[1],[].concat(r),Object.assign({},s)));case"fv":if(1!==l.length)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. fv(X) should have 1 arg.");return S.get(l[0],!1);default:throw new i.LambdaParseError("Malformed LMNtal Lambda Term. Unexpected atom name: "+o)}}(t,[],{});function e(t){if("("!==(t=t.trim())[0]||")"!==t[t.length-1])throw new i.LambdaParseError("Malformed LMNtal Lambda Term. Invalid Parentheses.");let e=0,n="",r=[];for(let i=1;i<t.length-1;i++){let s=t[i];"("===s?e++:")"===s&&e--,0===e&&","===s?(r.push(n),n=""):n+=s}if(0!==e)throw new i.LambdaParseError("Malformed LMNtal Lambda Term. Invalid Parentheses.");if(r.push(n),1===r.length){let t=r[0].trim();return""===t?[]:[t]}let s=[];for(let t of r)s.push(t.trim());return s}},e.makeTerms=function(t){return function t(e,n){let r=[].concat(e);if(0===n)return r;let i=t(e,n-1);for(let t=0;t<i.length;t++)for(let e=0;e<i.length;e++)r.push(new L(i[t],i[e]));let s=x.getNew(e),a=t(e.concat(s),n-1);for(let t of a)r.push(new E(s,t));return r}([],t)};class h{constructor(t){this.left="",this.right="",this.texLeft="",this.texRight="",this.type=t}addLeft(t){this.left=t+this.left}addRight(t){this.right+=t}addTexLeft(t){this.texLeft=t+this.texLeft}addTexRight(t){this.texRight+=t}static makeNext(t,e,n,r,i,s){let a=[].concat(t);for(let t of a)t.next=s(t.next),t.addLeft(e),t.addRight(n),t.addTexLeft(r),t.addTexRight(i);return a}getName(){return this.type}getPos(){return this.left.length}static compare(t,e){let n=t.getPos(),r=e.getPos();return n===r?0:n-r}}e.Redex=h;class u extends h{constructor(t){super("beta"),this.content=t,this.la=t.left,this.next=this.la.expr.substitute(this.la.boundvar,t.right),this.arg=t.right,this.rule="beta"}toString(){let t=[],e=this.la.expr;for(;e instanceof E;)t.push(e.boundvar.toString(!1)),e=e.expr;let n=t.join("")+"."+e.toString(!0);return this.left+"(\\["+this.la.boundvar.toString(!1)+"]"+n+")["+this.arg.toString(!1)+"]"+this.right}toTexString(){let t=[],e=this.la.expr;for(;e instanceof E;)t.push(e.boundvar.toTexString(!1)),e=e.expr;let n=t.join("")+"."+e.toTexString(!0);return this.texLeft+"(\\strut \\lambda{\\underline{"+this.la.boundvar.toTexString(!1)+"}}"+n+")\\underline{\\strut "+this.arg.toTexString(!1)+"}"+this.texRight}toHTMLString(){let t=[],e=this.la.expr;for(;e instanceof E;)t.push(e.boundvar.toString(!1)),e=e.expr;let n=t.join("")+"."+e.toString(!0);return l(this.left)+'(\\<span class="lf-beta lf-boundvar">'+l(this.la.boundvar.toString(!1))+"</span>"+l(n)+')<span class="lf-beta lf-arg">'+l(this.arg.toString(!1))+"</span>"+l(this.right)}getTexRule(){return"\\beta"}}class c extends h{constructor(t){super("eta"),this.content=t,this.app=t.expr,this.next=this.app.left,this.rule="eta"}toString(){return this.left+"["+this.content+"]"+this.right}toTexString(){return this.texLeft+"\\underline{\\strut "+this.content.toTexString(this.left+this.right==="")+"}"+this.texRight}toHTMLString(){return l(this.left)+'<span class="lf-eta">('+l(this.content.toString(this.left+this.right===""))+")</span>"+l(this.right)}getTexRule(){return"\\eta"}}class d extends h{constructor(t){super("macro"),this.content=t,this.next=t.expr,this.rule="macro"}toString(){return this.left+"[<"+this.content.name+">]"+this.right}toTexString(){return this.texLeft+"\\underline{\\strut "+this.content.toTexString(!1)+"}"+this.texRight}toHTMLString(){return l(this.left)+'<span class="lf-macro">&lt;'+l(this.content.name)+"&gt;</span>"+l(this.right)}getTexRule(){return"{\\rm m}"}}class f extends h{constructor(t,e,n){super("typed"),this.content=t,this.next=e,this.rule=n}toString(){return this.left+"["+this.content.toString(this.left+this.right==="")+"]"+this.right}toTexString(){return this.texLeft+"\\underline{\\strut "+this.content.toTexString(!1)+"}"+this.texRight}toHTMLString(){return l(this.left)+'<span class="lf-typed">'+l(this.content.toString(this.left+this.right===""))+"</span>"+l(this.right)}getTexRule(){return"{\\rm ("+this.rule+")}"}}class p{constructor(t,e){this.eqs=t,this.proofTree=e}}class g{constructor(t){this.className=t}parseChurchNum(){if(!(this instanceof E))return null;const t=this.boundvar;let e=this.expr,n=0;if(!(e instanceof E))return null;const r=e.boundvar;for(e=e.expr;e instanceof L;){if(n++,!e.left.equals(t))return null;e=e.right}return e.equals(r)?n:null}parseChurchBool(){const t=new E(new x("x"),new E(new x("y"),new x("x"))),e=new E(new x("x"),new E(new x("y"),new x("y")));return!!this.equalsAlpha(t)||!this.equalsAlpha(e)&&null}toLMNtal(){throw new i.TypeError("Expression '"+this+"' cannot be converted into LMNtal (untyped only).")}}e.Expression=g;class m extends g{constructor(t,e){super(void 0===e?"Symbol":e),this.name=t}equals(t){return t instanceof m&&t.className===this.className&&t.name===this.name}equalsAlpha(t){return t instanceof m&&t.className===this.className&&t.name===this.name}toString(t){return this.name}toTexString(t){throw new i.TexError("class Symbol does not have tex string")}getFV(){return this.freevars}substitute(t,e){throw new i.SubstitutionError("Undefined Substitution")}getEquations(t,e,n){throw new i.TypeError("Undefined Type")}getRedexes(t,e,n){throw new i.ReductionError("Symbols must not appear in parsed Expression")}getLeftMostRedex(t,e,n){throw new i.ReductionError("Symbols must not appear in parsed Expression")}extractMacros(){throw new i.ReductionError("Symbols must not appear in parsed Expression")}}class x extends m{constructor(t){super(t,"Variable"),this.freevars=[this]}substitute(t,e){return this.equals(t)?e:this}getEquations(t,e,n){for(let n of t)if(n.equals(this)){let i="\\AxiomC{}\n";return i+="\\RightLabel{\\scriptsize(var)}\n",i+="\\UnaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.name+" : "+e.toTexString()+" $}\n",new p([new r.TypeEquation(n.type,e)],i)}throw new i.TypeError("free variable is not allowed: "+this)}toTexString(t){return this.name}static union(t,e,n){if(void 0===n){let n=[];for(let e of t)n.push(e);for(let r of x.dif(e,t))n.push(r);return n}return x.union(x.union(t,e),n)}static dif(t,e){let n=[];for(let r of t)x.contains(e,r)||n.push(r);return n}static contains(t,e){for(let n of t)if(n.equals(e))return!0;return!1}static gammaToTexString(t){if(0===t.length)return"";let e=t[0].name+" : "+t[0].type.toTexString();for(let n=1;n<t.length;n++)e+=",~"+t[n].name+" : "+t[n].type.toTexString();return e}static getNew(t){let e="abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ".split("");for(let n of e){let e=new x(n);if(!x.contains(t,e))return e}throw new i.SubstitutionError("No more Variables available")}getRedexes(t,e,n){return[]}getLeftMostRedex(t,e,n){return null}extractMacros(){return this}toLMNtal(){return"fv("+this.name+")"}}class w extends m{constructor(t,e){super(t,e),this.freevars=[]}substitute(t,e){return this}getEquations(t,e,n){let i="\\AxiomC{}\n";return i+="\\RightLabel{\\scriptsize(con)}\n",i+="\\UnaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p([new r.TypeEquation(this.type,e)],i)}toString(t){return"["+this.name+"]"}toTexString(t){return this.name+"^{"+this.type.toTexString()+"}"}getRedexes(t,e,n){if(t)return[];throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className)}getLeftMostRedex(t,e,n){if(t)return null;throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className)}extractMacros(){return this}}class b extends w{constructor(t){super(t.toString(),"ConstInt"),this.value=t,this.type=new r.TypeInt}}class T extends w{constructor(t){super(t.toString(),"ConstBool"),this.value=t,this.type=new r.TypeBool}}class y extends w{constructor(t){switch(super(t,"ConstOp"),t){case"+":this.value=((t,e)=>new b(t.value+e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeInt));break;case"-":this.value=((t,e)=>new b(t.value-e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeInt));break;case"*":this.value=((t,e)=>new b(t.value*e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeInt));break;case"/":this.value=((t,e)=>{if(0===e.value)throw new i.ReductionError("Dividing by '0' is not allowed");return new b(Math.floor(t.value/e.value))}),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeInt));break;case"%":this.value=((t,e)=>{if(0===e.value)throw new i.ReductionError("Dividing by '0' is not allowed");return new b(t.value-4*Math.floor(t.value/e.value))}),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeInt));break;case"<":this.value=((t,e)=>new T(t.value<e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeBool));break;case">":this.value=((t,e)=>new T(t.value>e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeBool));break;case"<=":this.value=((t,e)=>new T(t.value<=e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeBool));break;case">=":this.value=((t,e)=>new T(t.value>=e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeBool));break;case"==":this.value=((t,e)=>new T(t.value==e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeBool));break;case"!=":this.value=((t,e)=>new T(t.value!=e.value)),this.type=new r.TypeFunc(new r.TypeInt,new r.TypeFunc(new r.TypeInt,new r.TypeBool));break;case"eq":case"eq":this.value=((t,e)=>new T(t.value==e.value)),this.type=new r.TypeFunc(new r.TypeBool,new r.TypeFunc(new r.TypeBool,new r.TypeBool));break;case"xor":this.value=((t,e)=>new T(t.value!=e.value)),this.type=new r.TypeFunc(new r.TypeBool,new r.TypeFunc(new r.TypeBool,new r.TypeBool));break;case"or":this.value=((t,e)=>new T(t.value||e.value)),this.type=new r.TypeFunc(new r.TypeBool,new r.TypeFunc(new r.TypeBool,new r.TypeBool));break;case"and":this.value=((t,e)=>new T(t.value&&e.value)),this.type=new r.TypeFunc(new r.TypeBool,new r.TypeFunc(new r.TypeBool,new r.TypeBool));break;case"nor":this.value=((t,e)=>new T(!(t.value||e.value))),this.type=new r.TypeFunc(new r.TypeBool,new r.TypeFunc(new r.TypeBool,new r.TypeBool));break;case"nand":this.value=((t,e)=>new T(!(t.value&&e.value))),this.type=new r.TypeFunc(new r.TypeBool,new r.TypeFunc(new r.TypeBool,new r.TypeBool));break;default:throw new i.LambdaParseError("Undefined function: "+t)}}}class v extends m{substitute(t,e){return this}constructor(){super("nil","Nil"),this.freevars=[]}getEquations(t,e,n){let i=r.TypeVariable.getNew(),s=new r.TypeList(i),a="\\AxiomC{}\n";return a+="\\RightLabel{\\scriptsize(nil)}\n",a+="\\UnaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.name+" : "+s.toTexString()+" $}\n",new p([new r.TypeEquation(e,s)],a)}toString(t){return"["+this.name+"]"}toTexString(t){return"{\\rm "+this.name+"}"}getRedexes(t,e,n){if(t)return[];throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className)}getLeftMostRedex(t,e,n){if(t)return null;throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className)}extractMacros(){return this}}class S extends m{constructor(t,e,n,r){super(t,"Macro"),this.freevars=[],this.expr=e,this.typed=n,this.type=r}static add(t,e,n){let r=e.expr;if(!/^[a-zA-Z0-9!?]+$/.test(t))throw new i.MacroError("<"+t+"> cannot be used as a name of macro. Available characters: [a-zA-Z0-9!?]");let s=n?S.map:S.mapUntyped;if(0!==r.getFV().length)throw new i.MacroError("<"+t+"> contains free variables: "+r.getFV());let a=new S(t,r,n,e.type);return s[t]=a,s[t]}static clear(t){t?S.map={}:S.mapUntyped={}}static get(t,e){let n;return void 0===(n=e?S.map[t]:S.mapUntyped[t])?null!==t.match(/^\d+$/)?new S(t,function(t,e){let n=0===t?"z":"sz";for(let e=1;e<t;e++)n="s("+n+")";return o("\\sz."+n,e)}(parseInt(t),e),e,void 0):new S(t,"true"==t?o("\\xy.x",e):"false"==t?o("\\xy.y",e):void 0,e,void 0):new S(t,n.expr,e,n.type)}static getMap(t){return Object.assign({},t?S.map:S.mapUntyped)}substitute(t,e){return this}toString(t){return"<"+this.name+">"}equalsAlpha(t){return void 0===this.expr?this.equals(t):this.expr.equalsAlpha(t)}getEquations(t,e,n){if(void 0===this.expr)throw new i.TypeError(this+" is undefined.");return this.expr.getEquations(t,e,n)}toTexString(t){return"\\,\\overline{\\bf "+this.name+"}\\,"}getRedexes(t,e,n){let r=S.get(this.name,t);return void 0===r.expr?[]:[new d(r)]}getLeftMostRedex(t,e,n){let r=S.get(this.name,t);return void 0===r.expr?null:new d(r)}extractMacros(){return void 0===this.expr?this:this.expr.extractMacros()}toLMNtal(){return void 0===this.expr?"fv("+this.name+")":this.expr.toLMNtal()}}S.map={},S.mapUntyped={},e.Macro=S;class E extends g{constructor(t,e){super("LambdaAbstraction"),this.freevars=void 0,this.boundvar=t,this.expr=e}static parse(t,e){let n=[];for(;t.length>0;){let r=t.shift();if("."===r.name){let r=a(t,e);for(;n.length>0;)r=new E(n.pop(),r);return r}if(null===r.name.match(/^[A-Za-z]$/))throw new i.LambdaParseError("Unexpected token: '"+r+"'");n.push(new x(r.name))}throw new i.LambdaParseError("'.' is needed")}toString(t){let e=[this.boundvar],n=this.expr;for(;n instanceof E;)e.push(n.boundvar),n=n.expr;let r="\\"+e.join("")+"."+n.toString(!0);return t||(r="("+r+")"),r}getFV(){return void 0!==this.freevars?this.freevars:(this.freevars=[],this.freevars=x.dif(this.expr.getFV(),[this.boundvar]))}substitute(t,e){if(this.boundvar.equals(t))return this;if(x.contains(e.getFV(),this.boundvar)){let n=x.union(this.expr.getFV(),e.getFV()),r=x.getNew(n);return new E(r,this.expr.substitute(this.boundvar,r)).substitute(t,e)}return new E(this.boundvar,this.expr.substitute(t,e))}equals(t){return t instanceof E&&t.boundvar.equals(this.boundvar)&&t.expr.equals(this.expr)}equalsAlpha(t){if(!(t instanceof E))return!1;if(this.equals(t))return!0;let e=this.boundvar,n=this.expr,r=t.boundvar,i=t.expr;return x.contains(n.getFV(),r)?i.equalsAlpha(n):i.equalsAlpha(n.substitute(e,r))}getEquations(t,e,n){let i=r.TypeVariable.getNew(),s=r.TypeVariable.getNew();this.boundvar.type=s;let a=this.expr.getEquations(t.concat(this.boundvar),i,!0),o=a.proofTree;return o+="\\RightLabel{\\scriptsize(abs)}\n",o+="\\UnaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p(a.eqs.concat(new r.TypeEquation(e,new r.TypeFunc(s,i))),o)}toTexString(t){let e=[this.boundvar.toTexString(!1)],n=this.expr;for(;n instanceof E;)e.push(n.boundvar.toTexString(!1)),n=n.expr;let r="\\lambda "+e.join("")+"."+n.toTexString(!0);return t||(r="("+r+")"),r}isEtaRedex(){return this.expr instanceof L&&this.expr.right.equals(this.boundvar)&&!x.contains(this.expr.left.getFV(),this.boundvar)}getRedexes(t,e,n){if(t)return[];let r=[this.boundvar],i=this.expr;for(;i instanceof E;)r.push(i.boundvar),i=i.expr;let s="",a="";n||(s="(",a=")");let o=h.makeNext(i.getRedexes(!1,e,!0),s+"\\"+r.join("")+".",a,s+"\\lambda{"+r.join("")+"}.",a,t=>{let e=[].concat(r),n=t;for(;e.length>0;){let t=e.pop();n=new E(t,n)}return n});return void 0===e&&(console.error("etaAllowed is undefined."),e=!1),e&&this.isEtaRedex()&&o.push(new c(this)),o}getLeftMostRedex(t,e,n){if(t)return null;if(void 0===e&&(console.error("etaAllowed is undefined."),e=!1),e&&this.isEtaRedex())return new c(this);let r=[this.boundvar],i=this.expr;for(;i instanceof E;)r.push(i.boundvar),i=i.expr;let s="",a="";n||(s="(",a=")");let o=i.getLeftMostRedex(!1,e,!0);return null===o?null:(o.next=(t=>{let e=[].concat(r),n=t;for(;e.length>0;){let t=e.pop();n=new E(t,n)}return n})(o.next),o.addLeft(s+"\\"+r.join("")+"."),o.addRight(a),o.addTexLeft(s+"\\lambda{"+r.join("")+"}."),o.addTexRight(a),o)}extractMacros(){return new E(this.boundvar,this.expr.extractMacros())}toLMNtal(){let t=this.expr.toLMNtal().split("fv("+this.boundvar.name+")"),e=t[0],n=[];for(let r=1;r<t.length;r++){let i=s.LambdaFriends.getNewLink();n.push(i),e+=i+t[r]}return"lambda("+function t(e){switch(e.length){case 0:return"rm";case 1:return e[0];case 2:return"cp("+e[0]+","+e[1]+")";default:return"cp("+e.shift()+","+t(e)+")"}}(n)+","+e+")"}}class L extends g{constructor(t,e){super("Application"),this.left=t,this.right=e}isBetaRedex(){return this.left instanceof E}toString(t){let e=this.left.toString(this.left instanceof L)+this.right.toString(!1);return t||(e="("+e+")"),e}getFV(){return void 0===this.freevars?this.freevars=x.union(this.left.getFV(),this.right.getFV()):this.freevars}substitute(t,e){return new L(this.left.substitute(t,e),this.right.substitute(t,e))}equals(t){return t instanceof L&&t.left.equals(this.left)&&t.right.equals(this.right)}equalsAlpha(t){return t instanceof L&&t.left.equalsAlpha(this.left)&&t.right.equalsAlpha(this.right)}getEquations(t,e,n){let i=r.TypeVariable.getNew(),s=this.left.getEquations(t,new r.TypeFunc(i,e),!1),a=this.right.getEquations(t,i,!1),o=s.proofTree+a.proofTree;return o+="\\RightLabel{\\scriptsize(app)}\n",o+="\\BinaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p(s.eqs.concat(a.eqs),o)}toTexString(t){let e=this.left.toTexString(this.left instanceof L)+this.right.toTexString(!1);return t||(e="("+e+")"),e}getRedexes(t,e,n){if(t){let t=n?"":"(",e=n?"":")";if(this.left instanceof E)return[new f(this,this.left.expr.substitute(this.left.boundvar,this.right),"app2")];if(this.left instanceof L&&this.left.left instanceof y&&this.left.right instanceof w){let n=this.left.left,s=this.left.right,a=this.right;if(a instanceof w){if(n.type.left.equals(s.type)&&n.type.right instanceof r.TypeFunc&&n.type.right.left.equals(a.type))return[new f(this,n.value(s,a),"app5")];throw new i.ReductionError(n.type+" cannot handle "+s.type+" and "+a.type+" as arguments")}return h.makeNext(a.getRedexes(!0,!1,!1),t+n.toString(!1)+s.toString(!1),e,t+n.toTexString(!1)+s.toTexString(!1),e,t=>new L(new L(n,s),t))}return this.left instanceof y?h.makeNext(this.right.getRedexes(!0,!1,!1),t+this.left.toString(!1),e,t+this.left.toTexString(!1),e,t=>new L(this.left,t)):h.makeNext(this.left.getRedexes(!0,!1,!1),t,e+this.right.toString(!1),t,e+this.right.toTexString(!1),t=>new L(t,this.right))}{let t=[this],r=[""],i=[""];for(;;){let e=t[t.length-1];if(r.push(e.right.toString(!1)+r[r.length-1]),i.push(e.right.toTexString(!1)+r[i.length-1]),!(e.left instanceof L))break;t.push(e.left)}let s=t[t.length-1].left.getRedexes(!1,e,!1);for(;t.length>0;){let n=t.pop(),a=h.makeNext(s,"",n.right.toString(!1),"",n.right.toTexString(!1),t=>new L(t,n.right)),o=n.left.toString(!0),l=h.makeNext(n.right.getRedexes(!1,e,!1),o,"",n.left.toTexString(!1),"",t=>new L(n.left,t));s=a.concat(l),r.pop(),i.pop(),n.isBetaRedex()&&s.push(new u(n))}return n||(s=h.makeNext(s,"(",")","(",")",t=>t)),s}}getLeftMostRedex(t,e,n){if(t){let t=n?"":"(",e=n?"":")";if(this.left instanceof E)return new f(this,this.left.expr.substitute(this.left.boundvar,this.right),"app2");if(this.left instanceof L&&this.left.left instanceof y&&this.left.right instanceof w){let n=this.left.left,s=this.left.right,a=this.right;if(a instanceof w){if(n.type.left.equals(s.type)&&n.type.right instanceof r.TypeFunc&&n.type.right.left.equals(a.type))return new f(this,n.value(s,a),"app5");throw new i.ReductionError(n.type+" cannot handle "+s.type+" and "+a.type+" as arguments")}{let r=a.getLeftMostRedex(!0,!1,!1);return null===r?null:(r.next=new L(new L(n,s),r.next),r.addLeft(t+n.toString(!1)+s.toString(!1)),r.addRight(e),r.addTexLeft(t+n.toTexString(!1)+s.toTexString(!1)),r.addTexRight(e),r)}}if(this.left instanceof y){let n=this.right.getLeftMostRedex(!0,!1,!1);return null===n?null:(n.next=new L(this.left,n.next),n.addLeft(t+this.left.toString(!1)),n.addRight(e),n.addTexLeft(t+this.left.toTexString(!1)),n.addTexRight(e),n)}{let n=this.right.getLeftMostRedex(!0,!1,!1);return null===n?null:(n.next=new L(this.left,n.next),n.addLeft(t),n.addRight(e+this.right.toString(!1)),n.addTexLeft(t),n.addTexRight(e+this.right.toTexString(!1)),n)}}{let t=[this],r=[""],i=[""];for(;;){let e=t[t.length-1];if(r.push(e.right.toString(!1)+r[r.length-1]),i.push(e.right.toTexString(!1)+r[i.length-1]),!(e.left instanceof L))break;t.push(e.left)}let s=t[t.length-1].left.getLeftMostRedex(!1,e,!1);for(;t.length>0;){let n=t.pop();if(n.isBetaRedex())s=new u(n);else if(null===s){if(null!==(s=n.right.getLeftMostRedex(!1,e,!1))){let t=n.left.toString(!0);s.next=new L(n.left,s.next),s.addLeft(t),s.addRight(""),s.addTexLeft(n.left.toTexString(!1)),s.addTexRight("")}}else s.next=new L(s.next,n.right),s.addLeft(""),s.addRight(n.right.toString(!1)),s.addTexLeft(""),s.addTexRight(n.right.toTexString(!1));r.pop(),i.pop()}return null===s?null:(n||(s.addLeft("("),s.addRight(")"),s.addTexLeft("("),s.addTexRight(")")),s)}}extractMacros(){return new L(this.left.extractMacros(),this.right.extractMacros())}toLMNtal(){return"apply("+this.left.toLMNtal()+","+this.right.toLMNtal()+")"}}class q extends g{constructor(t,e){super("List"),this.head=t,this.tail=e}toString(t){let e=this.head.toString(!1)+"::"+this.tail.toString(!0);return t||(e="("+e+")"),e}getFV(){return void 0===this.freevars?this.freevars=x.union(this.head.getFV(),this.tail.getFV()):this.freevars}substitute(t,e){return new q(this.head.substitute(t,e),this.tail.substitute(t,e))}equals(t){return t instanceof q&&t.head.equals(this.head)&&t.tail.equals(this.tail)}equalsAlpha(t){return t instanceof q&&t.head.equalsAlpha(this.head)&&t.tail.equalsAlpha(this.tail)}getEquations(t,e,n){let i=r.TypeVariable.getNew(),s=new r.TypeList(i),a=this.head.getEquations(t,i,!1),o=this.tail.getEquations(t,s,!1),l=a.proofTree+o.proofTree;return l+="\\RightLabel{\\scriptsize(list)}\n",l+="\\BinaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+s.toTexString()+" $}\n",new p(a.eqs.concat(o.eqs,new r.TypeEquation(s,e)),l)}toTexString(t){let e=this.head.toTexString(!1)+"::"+this.tail.toTexString(!0);return t||(e="("+e+")"),e}getRedexes(t,e,n){if(t)return[];throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className)}getLeftMostRedex(t,e,n){if(t)return null;throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className)}extractMacros(){return new q(this.head.extractMacros(),this.tail.extractMacros())}}class R extends g{constructor(t,e,n){super("If"),this.state=t,this.ifTrue=e,this.ifFalse=n}getFV(){return this.freevars=x.union(this.state.getFV(),this.ifTrue.getFV(),this.ifFalse.getFV())}toString(t){let e="[if]"+this.state.toString(!0)+"[then]"+this.ifTrue.toString(!0)+"[else]"+this.ifFalse.toString(!0);return t||(e="("+e+")"),e}substitute(t,e){return new R(this.state.substitute(t,e),this.ifTrue.substitute(t,e),this.ifFalse.substitute(t,e))}equals(t){return t instanceof R&&t.state.equals(this.state)&&t.ifTrue.equals(this.ifTrue)&&t.ifFalse.equals(this.ifFalse)}equalsAlpha(t){return t instanceof R&&t.state.equalsAlpha(this.state)&&t.ifTrue.equalsAlpha(this.ifTrue)&&t.ifFalse.equalsAlpha(this.ifFalse)}getEquations(t,e,n){let i=this.state.getEquations(t,new r.TypeBool,!0),s=this.ifTrue.getEquations(t,e,!0),a=this.ifFalse.getEquations(t,e,!0),o=i.proofTree+s.proofTree+a.proofTree;return o+="\\RightLabel{\\scriptsize(if)}\n",o+="\\TrinaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p(i.eqs.concat(s.eqs,a.eqs),o)}toTexString(t){let e="{\\bf if}~"+this.state.toTexString(!0)+"~{\\bf then}~"+this.ifTrue.toTexString(!0)+"~{\\bf else}~"+this.ifFalse.toTexString(!0);return t||(e="("+e+")"),e}getRedexes(t,e,n){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);return this.state instanceof T?this.state.value?[new f(this,this.ifTrue,"if2")]:[new f(this,this.ifFalse,"if3")]:h.makeNext(this.state.getRedexes(!0,!1,!1),"([if]","[then]"+this.ifTrue.toString(!0)+"[else]"+this.ifFalse.toString(!0)+")","({\\bf if}~","~{\\bf then}~"+this.ifTrue.toTexString(!0)+"~{\\bf else}~"+this.ifFalse.toTexString(!0)+")",t=>new R(t,this.ifTrue,this.ifFalse))}getLeftMostRedex(t,e,n){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);if(this.state instanceof T)return this.state.value?new f(this,this.ifTrue,"if2"):new f(this,this.ifFalse,"if3");{let t=this.state.getLeftMostRedex(!0,!1,!1);return null===t?null:(t.next=new R(t.next,this.ifTrue,this.ifFalse),t.addLeft("([if]"),t.addRight("[then]"+this.ifTrue.toString(!0)+"[else]"+this.ifFalse.toString(!0)+")"),t.addTexLeft("({\\bf if}~"),t.addTexRight("~{\\bf then}~"+this.ifTrue.toTexString(!0)+"~{\\bf else}~"+this.ifFalse.toTexString(!0)+")"),t)}}static parse(t,e){let n=[],r=0,s=0,o=0;for(;;){if(0==t.length)throw new i.LambdaParseError("Illegal If statement");let e=t.shift();switch(e.name){case"if":r++;break;case"then":s++;break;case"else":o++}if(r===o&&s===r+1)break;n.push(e)}let l=a(n,e),h=[];for(r=0,s=0,o=0;;){if(0==t.length)throw new i.LambdaParseError("Illegal If statement");let e=t.shift();switch(e.name){case"if":r++;break;case"then":s++;break;case"else":o++}if(r===s&&o===r+1)break;h.push(e)}let u=a(h,e),c=a(t,e);return new R(l,u,c)}extractMacros(){return new R(this.state.extractMacros(),this.ifTrue.extractMacros(),this.ifFalse.extractMacros())}}class M extends g{constructor(t,e,n){super("Let"),this.boundvar=t,this.left=e,this.right=n}getFV(){if(void 0!==this.freevars)return this.freevars;let t=[];for(let e of this.right.getFV())e.equals(this.boundvar)||t.push(e);return this.freevars=x.union(t,this.left.getFV())}toString(t){let e="[let]"+this.boundvar.toString(!0)+"="+this.left.toString(!0)+"[in]"+this.right.toString(!0);return t||(e="("+e+")"),e}substitute(t,e){let n=this.left.substitute(t,e);if(this.boundvar.equals(t))return new M(this.boundvar,n,this.right);if(x.contains(e.getFV(),this.boundvar)){let r=x.union(this.right.getFV(),e.getFV()),i=x.getNew(r);return i.equals(t)?new M(i,n,this.right.substitute(this.boundvar,i)):new M(i,n,this.right.substitute(this.boundvar,i).substitute(t,e))}return new M(this.boundvar,n,this.right.substitute(t,e))}equals(t){return t instanceof M&&t.boundvar.equals(this.boundvar)&&t.left.equals(this.left)&&t.right.equals(this.right)}equalsAlpha(t){if(!(t instanceof M))return!1;if(this.equals(t))return!0;let e=this.boundvar,n=this.right,r=t.boundvar,i=t.right;return!x.contains(n.getFV(),r)&&i.equalsAlpha(n.substitute(e,r))}getEquations(t,e,n){let i=r.TypeVariable.getNew();this.boundvar.type=i;let s=this.left.getEquations(t,i,!0),a=this.right.getEquations(t.concat(this.boundvar),e,!0),o=s.proofTree+a.proofTree;return o+="\\RightLabel{\\scriptsize(let)}\n",o+="\\TrinaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p(s.eqs.concat(a.eqs),o)}toTexString(t){let e="{\\bf let}~"+this.boundvar.toTexString(!1)+" = "+this.left.toTexString(!0)+"~{\\bf in}~"+this.right.toTexString(!0);return t||(e="("+e+")"),e}getRedexes(t,e,n){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);return[new f(this,this.right.substitute(this.boundvar,this.left),"let")]}getLeftMostRedex(t,e,n){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);return new f(this,this.right.substitute(this.boundvar,this.left),"let")}static parse(t,e){let n=t.shift();if(null===n.name.match(/^[A-Za-z]$/))throw new i.LambdaParseError("Unexpected token: '"+n+"'");let r=new x(n.name);if("="!==t.shift().name)throw new i.LambdaParseError("'=' is expected");let s=[],o=1;for(;;){if(0==t.length)throw new i.LambdaParseError("Illegal Let statement");let e=t.shift();if("let"===e.name?o++:"in"===e.name&&o--,0==o)break;s.push(e)}let l=a(s,e),h=a(t,e);return new M(r,l,h)}extractMacros(){return new M(this.boundvar,this.left.extractMacros(),this.right.extractMacros())}}class N extends g{constructor(t,e,n,r,i){super("Case"),this.state=t,this.ifNil=e,this.head=n,this.tail=r,this.ifElse=i}getFV(){return void 0!==this.freevars?this.freevars:x.union(this.state.getFV(),this.ifNil.getFV(),x.dif(this.ifElse.getFV(),[this.head,this.tail]))}toString(t){let e="[case]"+this.state.toString(!0)+"[of][nil]->"+this.ifNil.toString(!0)+"|"+this.head.toString(!0)+"::"+this.tail.toString(!0)+"->"+this.ifElse.toString(!0);return t||(e="("+e+")"),e}substitute(t,e){let n=this.state.substitute(t,e),r=this.ifNil.substitute(t,e);if(this.head.equals(t)||this.tail.equals(t))return new N(n,r,this.head,this.tail,this.ifElse);if(x.contains(e.getFV(),this.head)||x.contains(e.getFV(),this.tail)){let i=this.head,s=this.tail,a=this.ifElse;if(x.contains(e.getFV(),i)){let n=x.union(this.ifElse.getFV(),e.getFV()),r=x.getNew(n);a=r.equals(t)?a.substitute(i,r):a.substitute(i,r).substitute(t,e),i=r}if(x.contains(e.getFV(),s)){let n=x.union(this.ifElse.getFV(),e.getFV()),r=x.getNew(n);a=r.equals(t)?a.substitute(s,r):a.substitute(s,r).substitute(t,e),s=r}return new N(n,r,i,s,a)}return new N(n,r,this.head,this.tail,this.ifElse.substitute(t,e))}equals(t){return t instanceof N&&t.state.equals(this.state)&&t.ifNil.equals(this.ifNil)&&t.head.equals(this.head)&&t.tail.equals(this.tail)&&t.ifElse.equals(this.ifElse)}equalsAlpha(t){return t instanceof N&&t.state.equalsAlpha(this.state)&&t.ifNil.equalsAlpha(this.ifNil)&&t.head.equalsAlpha(this.head)&&t.tail.equalsAlpha(this.tail)&&t.ifElse.equalsAlpha(this.ifElse)}getEquations(t,e,n){let i=r.TypeVariable.getNew(),s=new r.TypeList(i);this.head.type=i,this.tail.type=s;let a=this.state.getEquations(t,s,!0),o=this.ifNil.getEquations(t,e,!0),l=this.ifElse.getEquations(t.concat(this.head,this.tail),e,!0),h=a.proofTree+o.proofTree+l.proofTree;return h+="\\RightLabel{\\scriptsize(case)}\n",h+="\\TrinaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p(a.eqs.concat(o.eqs,l.eqs),h)}toTexString(t){let e="{\\bf case} "+this.state.toTexString(!0)+" {\\bf of} {\\rm nil} \\Rightarrow "+this.ifNil.toTexString(!0)+" | "+this.head.toTexString(!0)+"::"+this.tail.toTexString(!0)+" \\Rightarrow "+this.ifElse.toTexString(!0);return t||(e="("+e+")"),e}getRedexes(t,e){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);return this.state instanceof v?[new f(this,this.ifNil,"case2")]:this.state instanceof q?[new f(this,this.ifElse.substitute(this.head,this.state.head).substitute(this.tail,this.state.tail),"case3")]:h.makeNext(this.state.getRedexes(!0,!1,!0),"([case]","[of][nil]->"+this.ifNil+" | "+this.head+"::"+this.tail+"->"+this.ifElse+")","({\\bf case} "," {\\bf of} {\\rm nil} \\Rightarrow "+this.ifNil.toTexString(!0)+" | "+this.head.toTexString(!0)+"::"+this.tail.toTexString(!0)+" \\Rightarrow "+this.ifElse.toTexString(!0)+")",t=>new N(t,this.ifNil,this.head,this.tail,this.ifElse))}getLeftMostRedex(t,e){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);if(this.state instanceof v)return new f(this,this.ifNil,"case2");if(this.state instanceof q)return new f(this,this.ifElse.substitute(this.head,this.state.head).substitute(this.tail,this.state.tail),"case3");{let t=this.state.getLeftMostRedex(!0,!1,!0);return null===t?null:(t.next=new N(t.next,this.ifNil,this.head,this.tail,this.ifElse),t.addLeft("([case]"),t.addRight("[of][nil]->"+this.ifNil+" | "+this.head+"::"+this.tail+"->"+this.ifElse+")"),t.addTexLeft("({\\bf case} "),t.addTexRight(" {\\bf of} {\\rm nil} \\Rightarrow "+this.ifNil.toTexString(!0)+" | "+this.head.toTexString(!0)+"::"+this.tail.toTexString(!0)+" \\Rightarrow "+this.ifElse.toTexString(!0)+")"),t)}}static parse(t,e){let n=[],r=1;for(;;){if(0==t.length)throw new i.LambdaParseError("Illegal Case statement");let e=t.shift();if("case"===e.name?r++:"of"===e.name&&r--,0==r)break;n.push(e)}let s=a(n,e),o=t.shift();if("nil"!==o.name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");if("-"!==(o=t.shift()).name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");if(">"!==(o=t.shift()).name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");let l=[];for(r=1;;){if(0==t.length)throw new i.LambdaParseError("Too many [case]");let e=t.shift();if("case"===e.name?r++:"|"===e.name&&r--,0==r)break;l.push(e)}let h=a(l,e),u=new x(t.shift().name);if(null===u.name.match(/^[A-Za-z]$/))throw new i.LambdaParseError("Unexpected token: '"+u.name+"'");if(":"!==(o=t.shift()).name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");if(":"!==(o=t.shift()).name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");let c=new x(t.shift().name);if(null===c.name.match(/^[A-Za-z]$/))throw new i.LambdaParseError("Unexpected token: '"+c.name+"'");if("-"!==(o=t.shift()).name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");if(">"!==(o=t.shift()).name)throw new i.LambdaParseError("Unexpected token: '"+o+"'");let d=a(t,e);return new N(s,h,u,c,d)}extractMacros(){return new N(this.state.extractMacros(),this.ifNil.extractMacros(),this.head,this.tail,this.ifElse.extractMacros())}}class F extends g{constructor(t,e){super("Fix"),this.boundvar=t,this.expr=e}getFV(){if(void 0!==this.freevars)return this.freevars;let t=[];for(let e of this.expr.getFV())e.equals(this.boundvar)||t.push(e);return this.freevars=t}toString(t){let e="[fix]"+this.boundvar.toString(!1)+"."+this.expr.toString(!0);return t||(e="("+e+")"),e}substitute(t,e){if(this.boundvar.equals(t))return this;if(x.contains(e.getFV(),this.boundvar)){let n=x.union(this.expr.getFV(),e.getFV()),r=x.getNew(n);return new F(r,this.expr.substitute(this.boundvar,r)).substitute(t,e)}return new F(this.boundvar,this.expr.substitute(t,e))}equals(t){return t instanceof F&&t.boundvar.equals(this.boundvar)&&t.expr.equals(this.expr)}equalsAlpha(t){if(!(t instanceof F))return!1;if(this.equals(t))return!0;let e=this.boundvar,n=this.expr,r=t.boundvar,i=t.expr;return x.contains(n.getFV(),r)?i.equalsAlpha(n):i.equalsAlpha(n.substitute(e,r))}getEquations(t,e,n){this.boundvar.type=e;let r=this.expr.getEquations(t.concat(this.boundvar),e,!0),i=r.proofTree;return i+="\\RightLabel{\\scriptsize(fix)}\n",i+="\\UnaryInfC{$"+x.gammaToTexString(t)+" \\vdash "+this.toTexString(n)+" : "+e.toTexString()+" $}\n",new p(r.eqs,i)}toTexString(t){let e="{\\bf fix}~"+this.boundvar.toTexString(!0)+"."+this.expr.toTexString(!0);return t||(e="("+e+")"),e}getRedexes(t,e,n){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);return[new f(this,this.expr.substitute(this.boundvar,new F(new x(this.boundvar.name),this.expr)),"fix")]}getLeftMostRedex(t,e,n){if(!t)throw new i.ReductionError("Untyped Reduction cannot handle typeof "+this.className);return new f(this,this.expr.substitute(this.boundvar,new F(new x(this.boundvar.name),this.expr)),"fix")}static parse(t,e){let n=t.shift();if(null===n.name.match(/^[A-Za-z]$/))throw new i.LambdaParseError("Unexpected token: '"+n+"'");let r=new x(n.name);if("."!==t.shift().name)throw new i.LambdaParseError("'.' is expected");let s=a(t,e);return new F(r,s)}extractMacros(){return new F(this.boundvar,this.expr.extractMacros())}}},function(t,e,n){"use strict";Object.defineProperty(e,"__esModule",{value:!0});const r=n(1),i=n(0);class s{constructor(t,e){this.children=[],this.label=t,this.info=e,this.id=e.nextId++,e.nodes.push(this)}toString(){return this.label}equalsShape(t){return this.info.nodes.length===t.info.nodes.length&&this.info.edges.length===t.info.edges.length&&this.info.allowMultipleEdges===t.info.allowMultipleEdges&&function t(e,n,r){if(e.children.length!==n.children.length)return!1;r.push({n1:e,n2:n});let i=[].concat(n.children);let s=!0;for(let n of e.children){let e=!1;for(let s=0;s<i.length;s++){let a=i[s],o=void 0;for(let t of r)if(t.n1.id===n.id||t.n2.id===a.id){o=t;break}if(void 0===o&&t(n,a,r)||void 0!==o&&o.n1.id===n.id&&o.n2.id===a.id){i.splice(s,1),e=!0;break}}if(!e){s=!1;break}}for(let t=0;t<r.length;t++)if(r[t].n1.id===e.id)return r.splice(t,1),s;throw new Error("Unexpected Error")}(this,t,[])}static parse(t){let e=t.split(";"),n=new a([],[],null,null,0,!1),i={};for(let t of e){let e=t.split("->"),r=o(e[0]);for(let t=1;t<e.length;t++){let i=o(e[t]);for(let t of r)for(let e of i)n.hasEdge({from:t,to:e})&&(n.allowMultipleEdges=!0),n.edges.push({from:t,to:e}),t.children.push(e);r=i}}if(void 0===i.root)throw new r.GraphParseError("there must be a 'root' node.");return i.root;function o(t){let e=t.split(","),a=[];for(let t of e){if(""===(t=t.trim()))continue;if(!/^[a-z][a-z0-9]*$/.test(t))throw new r.GraphParseError("the name '"+t+"' cannot be used as node name, must be [a-z][a-z0-9]*");let e=i[t];void 0===e&&(e=i[t]=new s(t,n)),a.push(e)}return a}}static search(t,e){let r,s="graph_closure.csv";try{let t=new XMLHttpRequest;t.open("GET",s,!1),t.send(null),r=t.responseText}catch(t){0;let e=n(6);try{e.statSync(s)}catch(t){return void console.error("File Not Found: "+s)}r=e.readFileSync(s,"utf8")}let a=r.split("\n");for(let n of a){let r=n.split(",");if(parseInt(r[2])===t.info.nodes.length&&parseInt(r[3])===t.info.edges.length){let n=new i.LambdaFriends(r[0],!1,!1,e);for(;null!==n.deepen(););if(n.root.equalsShape(t))return n}}return null}}e.GraphNode=s;class a{constructor(t,e,n,r,i,s){this.nodes=t,this.edges=e,this.typed=n,this.etaAllowed=r,this.nextId=i,this.allowMultipleEdges=s}hasEdge(t){for(let e of this.edges)if(e.from.id===t.from.id&&e.to.id===t.to.id)return!0;return!1}}class o extends s{constructor(t,e,n){super((t=t.extractMacros()).toString(!0),n),this.expr=t,this.parent=e,this.depth=null===e?0:e.depth+1,this.nextrs=this.expr.getRedexes(n.typed,n.etaAllowed,!0),this.isNormalForm=0===this.nextrs.length}static makeRoot(t,e,n,r){return new o(t,null,new a([],[],e,n,0,r))}visit(){let t={nodes:[],edges:[]};for(let e of this.nextrs){let n=this.find(e.next);if(null===n){let n=new o(e.next,this,this.info);this.children.push(n),t.nodes.push(n),!this.info.allowMultipleEdges&&this.info.hasEdge({from:this,to:n})||(t.edges.push({from:this,to:n}),this.info.edges.push({from:this,to:n}))}else this.children.push(n),!this.info.allowMultipleEdges&&this.info.hasEdge({from:this,to:n})||(t.edges.push({from:this,to:n}),this.info.edges.push({from:this,to:n}))}return t}find(t){for(let e of this.info.nodes)if(e.expr.equalsAlpha(t))return e;return null}}e.ReductionNode=o},function(t,e){t.exports=require("fs")}]);
