+++ run.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+cd `dirname $0`
+make > /dev/null
+node js/cli.js $*
